% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add_node_attributes}
\alias{add_edge_attributes}
\alias{copy_node_attributes}
\alias{mutate_edges}
\title{Adding and copying attributes from one graph to another}
\usage{
add_node_attributes(object, attr_name, vector)

add_edge_attributes(object, object2)

copy_node_attributes(object, object2)

mutate_edges(data, object, name)
}
\arguments{
\item{object}{A data frame edgelist, matrix, igraph, tidygraph, or
network object.}

\item{attr_name}{Name of the new attribute in the resulting object.}

\item{vector}{A vector of values for the new attribute.}

\item{object2}{A second object to copy nodes or edges from.}
}
\description{
Adding and copying attributes from one graph to another
}
\examples{
mutate_edges(acmeEmails, sameSex, "sameSex")
}
