% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adolescent_society.R
\docType{data}
\name{adolescent_society}
\alias{adolescent_society}
\title{One-mode subset of adolescent society dataset}
\format{
tidygraph graph object
}
\usage{
data(adolescent_society)
}
\description{
One-mode subset of adolescent society dataset
}
\references{
Coleman, James S. 1961. The Adolescent Society.
New York:Free Press.

Feld, Scott. 1991. “Why your friends have more friends than you do”
American Journal of Sociology 96(6): 1464-1477.
}
\keyword{datasets}
