% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utilities.R
\name{arrange.vars}
\alias{arrange.vars}
\title{arrange.vars}
\source{
https://stackoverflow.com/questions/5620885/how-does-one-reorder-columns-in-a-data-frame
}
\usage{
arrange.vars(data, vars)
}
\arguments{
\item{data}{must be a data frame}

\item{vars}{must be a named vector, e.g. c("var.name"=1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Arranges data frame variables by position
}
\details{
Credit goes to the user landroni on StackOverflow for this very useful function
}
\examples{
arrange.vars(data, c("Out"=2, "Files"=1, "Time"=4))
}
