% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograph.R
\name{autograph}
\alias{autograph}
\title{Autograph with sensible defaults}
\usage{
autograph(object, node_color = NULL, ...)
}
\arguments{
\item{object}{migraph-consistent object}

\item{node_color}{node variable in quotation marks
that should be used for colouring the nodes}
}
\description{
Autograph with sensible defaults
}
