% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autographr.R
\name{autographr}
\alias{autographr}
\title{Quickly graph networks with sensible defaults}
\usage{
autographr(object, algorithm = c("fr", "kk", "stress"), node_color = NULL, ...)
}
\arguments{
\item{object}{migraph-consistent object}

\item{algorithm}{an igraph layout algorithm,
currently defaults to Fruchterman-Reingold,
but Kamada-Kawai and 'stress' also available}

\item{node_color}{node variable in quotation marks
that should be used for colouring the nodes}

\item{...}{extra arguments}
}
\description{
The aim of this function is to provide users with a quick and easy
graphing function that makes best use of the data,
whatever its composition.
}
\examples{
autographr(adolescent_society)
autographr(ison_karateka)
}
