% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_blockmodel.R
\name{blockmodel_vis}
\alias{blockmodel_vis}
\alias{plot.block_model}
\title{ggplot2-based plotting of blockmodel results}
\usage{
\method{plot}{block_model}(x, ...)
}
\arguments{
\item{x}{A blockmodel-class object.}

\item{...}{Additional arguments passed on to ggplot2.}
}
\description{
ggplot2-based plotting of blockmodel results
}
\examples{
# Unsigned
usa_concor <- blockmodel_concor(mpn_elite_usa_advice)
plot(usa_concor)

# Signed network (Needs sign)
# marvel_concor <- blockmodel_concor(ison_marvel_relationships)
# plot(marvel)
}
