% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{cluster_triad_census}
\alias{cluster_triad_census}
\title{Get triad census summary by group
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
cluster_triad_census(object, clusters)
}
\value{
\code{group_triad_census()}
}
\description{
Deprecated on 2021-10-26.
}
\keyword{internal}
