% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.R
\name{connectedness}
\alias{connectedness}
\alias{node_components}
\alias{graph_components}
\alias{node_cuts}
\alias{edge_bridges}
\title{Network connectedness}
\usage{
node_components(object, method = c("weak", "strong"))

graph_components(object, method = c("weak", "strong"))

node_cuts(object)

edge_bridges(object)
}
\arguments{
\item{object}{a migraph-consistent object}

\item{method}{For directed networks,
either \code{weak} if edge direction is irrelevant,
or \code{strong} if edge direction is salient.
Ignored if network undirected.}
}
\description{
Network connectedness
}
\section{Functions}{
\itemize{
\item \code{node_components}: Returns nodes' component membership

\item \code{graph_components}: Returns number of components in the network

\item \code{node_cuts}: Returns logical of which nodes cut
or act as articulation points in a network

\item \code{edge_bridges}: Returns logical of which nodes cut
or act as articulation points in a network
}}

