% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_lattice}
\alias{create_lattice}
\title{Two-mode lattice}
\usage{
create_lattice(node1, node2)
}
\arguments{
\item{node1}{Number of nodes in the first node set}

\item{node2}{Number of nodes in the second node set}
}
\description{
Creates a two-mode lattice
}
\details{
Will construct a bilateral lattice,
with two ties for every second-mode node.
}
\examples{
\dontrun{
create_lattice(10, 12)
}
}
