% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_match}
\alias{create_match}
\title{Matched two-mode graph}
\usage{
create_match(node1, node2)
}
\arguments{
\item{node1}{Number of nodes in the first node set}

\item{node2}{Number of nodes in the second node set}
}
\description{
Creates a matched two-mode network
}
\details{
Will construct an affiliation matrix,
with by default both node1 and node2 matched.
TODO: Make node set matching optional.
}
\examples{
\dontrun{
create_match(10, 12)
}
}
