% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_nest}
\alias{create_nest}
\title{Nested two-mode graph}
\usage{
create_nest(n1, n2, as = c("tbl_graph", "igraph", "matrix"))
}
\arguments{
\item{n1}{Number of nodes in the first node set}

\item{n2}{Number of nodes in the second node set}

\item{as}{What type of object to return.}
}
\description{
Creates a nested two-mode network
}
\details{
Will construct an affiliation matrix,
with decreasing fill across n2.
}
\examples{
\dontrun{
create_nest(10, 12)
}
}
