% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_poles}
\alias{create_poles}
\title{Two-component two-mode graph}
\usage{
create_poles(node1, node2)
}
\arguments{
\item{node1}{Number of nodes in the first node set}

\item{node2}{Number of nodes in the second node set}
}
\description{
Creates a two-component two-mode network
}
\details{
Will construct an affiliation matrix,
with full component diagonal.
}
\examples{
\dontrun{
create_poles(10, 12)
}
}
