% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{create_random}
\alias{create_random}
\title{Two-mode random graph}
\usage{
create_random(node1, node2, density)
}
\arguments{
\item{node1}{Number of nodes in the first node set}

\item{node2}{Number of nodes in the second node set}

\item{density}{Number of edges in the network over the number of edges possible}
}
\description{
Creates a random two-mode network
}
\details{
Will construct an affiliation matrix,
with a random probability of a tie.
}
\examples{
\dontrun{
create_random(10, 12, 0.25)
}
}
