% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraph-defunct.R
\name{defunct}
\alias{defunct}
\alias{ggidentify}
\title{Functions that have been renamed, superseded, or are no longer working}
\usage{
ggidentify(object, node_measure, identify_function = max)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\section{Functions}{
\itemize{
\item \code{ggidentify}: Deprecated on 2022-01-24.
This function is deprecated and its functionality is included in the more
general purpose \code{autographr()} function. Please refer to its documentation
for more details about the new implementation.
}}

\keyword{internal}
