% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_diffusion.R
\name{diffusion}
\alias{diffusion}
\alias{network_transmissibility}
\alias{network_infection_length}
\alias{network_reproduction}
\alias{node_adoption_time}
\alias{node_adopter}
\alias{node_thresholds}
\alias{node_infection_length}
\title{Functions to measure diffusion processes on networks}
\usage{
network_transmissibility(diff_model)

network_infection_length(diff_model)

network_reproduction(diff_model)

node_adoption_time(diff_model)

node_adopter(diff_model)

node_thresholds(diff_model)

node_infection_length(diff_model)
}
\arguments{
\item{diff_model}{A valid network diffusion model.}
}
\description{
Functions to measure diffusion processes on networks
}
\section{Functions}{
\itemize{
\item \code{network_transmissibility()}: Calculates the average transmissibility observed
in a diffusion simulation, or the number of new infections over
the number of susceptible, over the number of infected

\item \code{network_infection_length()}: Calculates the average length nodes remain
infected in a compartmental model with recovery for the network as a whole

\item \code{network_reproduction()}: Calculates the observed reproductive number
in a diffusion simulation as the network's transmissibility over
the network's average infection length

\item \code{node_adoption_time()}: Measures nodes' time of adoption/infection

\item \code{node_adopter()}: Classifies membership of nodes into diffusion categories

\item \code{node_thresholds()}: Measures nodes' thresholds from the amount
of exposure they had when they became infected

\item \code{node_infection_length()}: Measures the average length nodes that become
infected remain infected in a compartmental model with recovery

}}
\examples{
  smeg <- manynet::generate_smallworld(15, 0.025)
  smeg_diff <- play_diffusion(smeg, recovery = 0.05)
  plot(smeg_diff)
  (adopts <- node_adopter(smeg_diff))
  summary(adopts)
  summary(node_adoption_time(smeg_diff), membership = adopts)
  summary(node_thresholds(smeg_diff), membership = adopts)
  summary(node_infection_length(smeg_diff))
}
\references{
Kermack, W. and McKendrick, A., 1927. "A contribution to the mathematical theory of epidemics".
\emph{Proc. R. Soc. London A} 115: 700-721.

Valente, Tom W. (1995). \emph{Network models of the diffusion of innovations}
(2nd ed.). Cresskill N.J.: Hampton Press.
}
\seealso{
Other measures: 
\code{\link{between_centrality}},
\code{\link{close_centrality}},
\code{\link{closure}},
\code{\link{cohesion}()},
\code{\link{degree_centrality}},
\code{\link{eigenv_centrality}},
\code{\link{features}},
\code{\link{heterogeneity}},
\code{\link{hierarchy}},
\code{\link{holes}}
}
\concept{measures}
