% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_features.R
\name{features}
\alias{features}
\alias{graph_smallworld}
\alias{graph_balance}
\alias{graph_modularity}
\title{Network/topological features}
\source{
\code{{signnet}} by David Schoch
}
\usage{
graph_smallworld(object, times = 100)

graph_balance(object, method = "triangles")

graph_modularity(object, membership = NULL, resolution = 1)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix, from base R
\item edgelist, a data frame from base R or tibble from tibble
\item igraph, from the igraph package
\item network, from the network package
\item tbl_graph, from the tidygraph package
}}

\item{times}{Integer of number of simulations.}

\item{method}{One of "triangles" (the default), "walk", or "frustration".}

\item{membership}{A vector of partition membership}

\item{resolution}{A proportion indicating the resolution scale.
By default 1.}
}
\description{
Network/topological features
}
\section{Functions}{
\itemize{
\item \code{graph_smallworld}: Returns small-world metrics for one- and two-mode networks.

\item \code{graph_balance}: Returns the structural balance index on
the proportion of balanced triangles,
ranging between \code{0} if all triangles are imbalanced and
\code{1} if all triangles are balanced.

\item \code{graph_modularity}: Returns modularity of one- or two-mode networks
}}

\examples{
graph_smallworld(ison_southern_women)
graph_smallworld(ison_brandes)
graph_modularity(ison_adolescents, 
  node_kernighanlin(ison_adolescents))
graph_modularity(ison_southern_women, 
  node_kernighanlin(ison_southern_women))
}
\references{
Watts, Duncan J., and Steven H. Strogatz. 1998.
“Collective Dynamics of ‘Small-World’ Networks.”
\emph{Nature} 393(6684):440–42.

Murata, Tsuyoshi. 2010. Modularity for Bipartite Networks.
In: Memon, N., Xu, J., Hicks, D., Chen, H. (eds)
\emph{Data Mining for Social Network Data. Annals of Information Systems}, vol 12.
Springer, Boston, MA.
\doi{10.1007/978-1-4419-6287-4_7}
}
\seealso{
\code{\link{graph_transitivity}} and \code{\link{graph_equivalency}}
for how clustering is calculated

Other measures: 
\code{\link{centrality}},
\code{\link{cohesion}()},
\code{\link{connectedness}},
\code{\link{diversity}},
\code{\link{node_constraint}()}
}
\concept{measures}
