% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggevolution.R
\name{ggatyear}
\alias{ggatyear}
\title{Plotting network at particular timepoint (year)}
\usage{
ggatyear(edgelist, year)
}
\arguments{
\item{edgelist}{a manyverse edgelist, expecting \code{Beg} and \code{End} variables,
among others}

\item{year}{numeric year, gets expanded to first of January that year}
}
\description{
Plotting network at particular timepoint (year)
}
\examples{
\dontrun{
ggatyear(membs, 1900)
}
}
