% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggidentify.R
\name{ggidentify}
\alias{ggidentify}
\title{Visualising graphs and identifying nodes with maximum values of the specified
measure.}
\usage{
ggidentify(object, FUN)
}
\arguments{
\item{object}{a migraph-consistent object}

\item{FUN}{some arbitrary function that runs on the object and
returns a numeric vector that can be used to scale the nodes}
}
\description{
Visualising graphs and identifying nodes with maximum values of the specified
measure.
}
\examples{
ggidentify(brandes, node_degree)
ggidentify(brandes, node_betweenness)
ggidentify(brandes, node_closeness)
ggidentify(brandes, node_eigenvector)
}
