% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraphlabel.R
\name{ggraphlabel}
\alias{ggraphlabel}
\title{Plot graph with quick labels}
\usage{
ggraphlabel(x, algorithm = c("kk", "fr"))
}
\arguments{
\item{x}{A migraph-consistent network/graph}

\item{algorithm}{An initial network layout,
currently either Kamada-Kawai ("kk") or
Fruchterman-Reingold ("fr")}
}
\description{
For quick and easy graphing of networks with labels
}
\examples{
ggraphlabel(adolescent_society)
}
