% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggidentify.R
\name{ggtools}
\alias{ggtools}
\alias{ggidentify}
\alias{ggdistrib}
\title{Visualising graphs and identifying nodes with maximum values of the specified
measure.}
\usage{
ggidentify(object, node_measure, identify_function = max)

ggdistrib(object, node_measure)
}
\arguments{
\item{object}{a migraph-consistent object}

\item{node_measure}{some arbitrary function that runs on the object and
returns a numeric vector that can be used to scale the nodes}

\item{identify_function}{a function for the identification of a single node,
e.g. max, min, mean, etc.}
}
\description{
Visualising graphs and identifying nodes with maximum values of the specified
measure.
}
\examples{
ggidentify(brandes, node_degree)
ggidentify(brandes, node_betweenness)
ggidentify(brandes, node_closeness)
ggidentify(brandes, node_eigenvector)
ggdistrib(brandes, node_degree)
ggdistrib(brandes, node_betweenness)
ggdistrib(brandes, node_closeness)
ggdistrib(brandes, node_eigenvector)
}
