% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\alias{node_names}
\alias{node_mode}
\alias{node_attribute}
\alias{edge_attribute}
\alias{edge_weights}
\alias{edge_signs}
\alias{graph_nodes}
\alias{graph_edges}
\alias{graph_dims}
\alias{graph_node_attributes}
\alias{graph_edge_attributes}
\title{Helpers to grab various attributes from nodes or edges in a graph}
\usage{
node_names(object)

node_mode(object)

node_attribute(object, attribute)

edge_attribute(object, attribute)

edge_weights(object)

edge_signs(object)

graph_nodes(object)

graph_edges(object)

graph_dims(object)

graph_node_attributes(object)

graph_edge_attributes(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix, from base R
\item edgelist, a data frame from base R or tibble from tibble
\item igraph, from the igraph package
\item network, from the network package
\item tbl_graph, from the tidygraph package
}}

\item{attribute}{An attribute name.}
}
\description{
Helpers to grab various attributes from nodes or edges in a graph
}
