% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{graph_balance}
\alias{graph_balance}
\title{Structural balance}
\usage{
graph_balance(object, method = "triangles")
}
\arguments{
\item{object}{a migraph-consistent object}

\item{method}{one of "triangles" (the default), "walk", or "frustration".}
}
\value{
"triangles" returns the proportion of balanced triangles,
ranging between \code{0} if all triangles are imbalanced and \code{1} if all triangles are balanced.

"walk" is based on eigenvalues of the signed and underlying unsigned network.
Check the paper by Estrada for details.

“frustration” assumes that the network can be partitioned into two groups,
where intra group edges are positive and inter group edges are negative.
The index is defined as the sum of intra group negative and inter group positive edges.
Note that the problem is NP complete and only an upper bound is returned (based on simulated annealing).
Exact methods can be found in the work of Aref.
}
\description{
Structural balance
}
\references{
Estrada, Ernesto (2019). Rethinking structural balance in signed social networks.
\emph{Discrete Applied Mathematics}, 268: 70--90.

Aref, Samin, and Mark C Wilson (2018). Measuring partial balance in signed networks.
\emph{Journal of Complex Networks}, 6(4): 566–595.
}
