% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{graph_mixed_census}
\alias{graph_mixed_census}
\alias{graph_census}
\alias{graph_dyad_census}
\alias{graph_triad_census}
\title{Censuses for the whole graph}
\usage{
graph_mixed_census(object, object2, B2 = NULL)

graph_dyad_census(object)

graph_triad_census(object)
}
\arguments{
\item{object}{a migraph-consistent object}
}
\description{
Censuses for the whole graph
}
\examples{
(mixed_cen <- graph_mixed_census(to_unsigned(ison_marvel_relationships, "positive"), ison_marvel_teams))
graph_dyad_census(ison_coleman)
graph_triad_census(ison_coleman)
}
