% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{graph_triad_census}
\alias{graph_triad_census}
\title{Triad census for the whole graph}
\usage{
graph_triad_census(object)
}
\arguments{
\item{object}{a migraph-consistent object}
}
\description{
Triad census for the whole graph
}
\examples{
graph_triad_census(ison_coleman)
}
