% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggidentify.R
\name{identify_clusters}
\alias{identify_clusters}
\alias{ggtree}
\alias{ggidentify_clusters}
\title{Plots for deciding on the number of network clusters}
\usage{
ggtree(hc, k = NULL)

ggidentify_clusters(hc, mat, method = "elbow")
}
\arguments{
\item{hc}{a hierarchical cluster object}

\item{k}{number of clusters. By default NULL,
but, if specified, \code{ggtree} will color branches and
add a line to indicate where the corresponding cluster
cut would be.}

\item{mat}{the matrix}

\item{method}{only "elbow" is currently implemented.}
}
\description{
Plots for deciding on the number of network clusters
}
\examples{
res <- cluster_regular_equivalence(mpn_elite_mex)
ggtree(res, 4)
ggidentify_clusters(res, mpn_mex_elite)
}
