% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is}
\alias{is}
\alias{is_twomode}
\alias{is_weighted}
\alias{is_directed}
\alias{is_labelled}
\alias{is_signed}
\alias{is_connected}
\alias{is_complex}
\alias{is_migraph}
\alias{is_graph}
\alias{is_chordal}
\alias{is_dag}
\alias{is_degseq}
\alias{is_graphical}
\alias{is_matching}
\alias{is_max_matching}
\alias{any_multiple}
\alias{which_multiple}
\alias{which_mutual}
\alias{is_simple}
\title{Tests of network properties}
\usage{
is_twomode(object)

is_weighted(object)

is_directed(object)

is_labelled(object)

is_signed(object)

is_connected(object, method = c("weak", "strong"))

is_complex(object)

is_migraph(object)

is_graph(object)

is_chordal(object, ...)

is_dag(object)

is_degseq(object, ...)

is_graphical(object, ...)

is_matching(object, matching, types = NULL)

is_max_matching(object, matching, types = NULL)

any_multiple(object)

which_multiple(object)

which_multiple(object)

which_mutual(object)

is_simple(object)
}
\arguments{
\item{object}{A migraph-consistent class object
(matrix, edgelist, igraph, network, tidygraph)}

\item{method}{Whether to identify components if only "weak"ly connected
or also "strong"ly connected.}

\item{...}{Additional parameters passed onto the underlying \code{is_} function.}
}
\value{
TRUE if object is a two-mode network, otherwise FALSE

TRUE if object is a weighted network, otherwise FALSE

TRUE if object is a directed network, otherwise FALSE

TRUE if object is a labelled network, otherwise FALSE

TRUE if object is a signed network, otherwise FALSE

TRUE if object is a connected network, otherwise FALSE
}
\description{
Tests of network properties
}
\examples{
is_twomode(southern_women)
is_weighted(southern_women)
is_directed(southern_women)
is_labelled(southern_women)
is_signed(southern_women)
is_connected(southern_women)
}
\seealso{
Other manipulation: 
\code{\link{coercion}},
\code{\link{project}},
\code{\link{to}}
}
\concept{manipulation}
