% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_adolescents}
\alias{ison_adolescents}
\title{One-mode subset of the adolescent society (Coleman 1961)}
\format{
A undirected one-mode tbl_graph object of 8 named nodes and 10
edges.
}
\usage{
data(ison_adolescents)
}
\description{
One-mode subset of the adolescent society (Coleman 1961)
}
\references{
Coleman, James S. 1961. The Adolescent Society.
New York:Free Press.

Feld, Scott. 1991. “Why your friends have more friends than you do”
American Journal of Sociology 96(6): 1464-1477.
}
\keyword{datasets}
