% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_algebra}
\alias{ison_algebra}
\title{Multiplex graph object of friends, social, and task ties (McFarland 2001)}
\format{
Multiplex \code{tbl_graph} object of friends, social, and task ties
between 16 anonymous students.
}
\source{
See also \code{data(studentnets.M182, package = "NetData")}
Larger comprehensive data set publicly available, contact Daniel A. McFarland for details.
}
\usage{
data(ison_algebra)
}
\description{
Multiplex graph object of friends, social, and task ties (McFarland 2001)
}
\details{
Multiplex graph object of friends, social, and task ties between 16 #' anonymous students.
M182 was an honors algebra class where researchers
collected friendship, social, and task ties between 16 students.
The edge attribute \code{friend_ties} contains friendship ties,
where \code{2} = best friends, \code{1} = friend, and \code{0} is not a friend.
\code{social_ties} consists of social interactions per hour,
and \code{task_ties} consists of task interactions per hour.
}
\references{
McFarland, Daniel A. (2001) ``Student Resistance.''
American Journal of Sociology, 107(3), p 612-678.
}
\keyword{datasets}
