% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_brandes}
\alias{ison_brandes}
\title{One-mode centrality demonstration network}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tbl_graph: 11 nodes and 12 edges
## #
## # An undirected simple graph with 1 component
## #
## # Node Data: 11 × 0 (active)
## # … with 5 more rows
## #
## # Edge Data: 12 × 2
##    from    to
##   <int> <int>
## 1     1     3
## 2     2     3
## 3     3     4
## # … with 9 more rows
}\if{html}{\out{</div>}}
}
\usage{
data(ison_brandes)
}
\description{
This network should solely be used
for demonstration purposes as it does not describe a real network.
}
\keyword{datasets}
