% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_community}
\alias{ison_community}
\alias{ison_karateka}
\title{Zachary's kareteka network}
\format{
Undirected one-mode igraph with 34 nodes and 78 edges
}
\usage{
data(ison_karateka)
}
\description{
Zachary's kareteka network
}
\keyword{datasets}
