% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_eies}
\alias{ison_eies}
\title{1DW 32(2) 440(1) EIES messages}
\format{
tidygraph graph object
}
\source{
networkdata package
}
\usage{
data(ison_eies)
}
\description{
1DW 32(2) 440(1) EIES messages
}
\references{
Freeman, S. C. and L. C. Freeman (1979).
The networkers network: A study of the impact of a new communications medium on sociometric structure.
Social Science Research Reports No 46. Irvine CA, University of California.

Wasserman S. and K. Faust (1994).
Social Network Analysis: Methods and Applications.
Cambridge University Press, Cambridge.
}
\keyword{datasets}
