% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_karateka}
\alias{ison_karateka}
\title{One-mode karateka network (Zachary 1977)}
\format{
Undirected one-mode igraph with 34 named nodes and 78 edges.
}
\usage{
data(ison_karateka)
}
\description{
One-mode karateka network (Zachary 1977)
}
\details{
Zachary's karateka network. The network was observed in a
university Karate club in 1977. The network describes association patterns
among 34 members and maps out allegiance patterns between members and either
Mr. Hi, the instructor, or the John A. the club president after an argument
about hiking the price for lessons. The allegiance of each node is
listed in the \code{obc} argument which takes the value 1 if the individual
sided with Mr. Hi after the fight and 2 if the individual sided with John A.
}
\references{
Zachary, Wayne W. 1977. “An Information Flow Model for Conflict and Fission in Small Groups.”
Journal of Anthropological Research 33(4):452–73.
}
\keyword{datasets}
