% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_lotr}
\alias{ison_lotr}
\title{One-mode interaction network of Lord of the Rings (book) character interactions}
\format{
One-mode tidygraph of 36 Lord of the Rings book characters and
66 interactions
}
\usage{
data(ison_lotr)
}
\description{
One-mode interaction network of Lord of the Rings (book) character interactions
}
\keyword{datasets}
