% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_marvel}
\alias{ison_marvel}
\alias{ison_marvel_teams}
\alias{ison_marvel_relationships}
\title{Multilevel two-mode affiliation, signed one-mode networks of Marvel comic
book characters (Y端ksel 2017)}
\format{
Two-mode igraph of 53 Marvel comic book characters and 141 team-ups,
with 683 team affiliations between them

One-mode igraph of 53 Marvel comic book characters and
558 signed (\code{1} = friends, \code{-1} = enemies) undirected ties
}
\source{
Umut Y端ksel, 31 March 2017
}
\usage{
data(ison_marvel_teams)

data(ison_marvel_relationships)
}
\description{
Multilevel two-mode affiliation, signed one-mode networks of Marvel comic
book characters (Y端ksel 2017)
}
\details{
This package includes two datasets related to the Marvel \emph{comic book} universe.
The first, \code{ison_marvel_teams},  is a two-mode affiliation network of 53
Marvel comic book characters and their affiliations to 141 different teams.
This network includes only information about nodes' names and nodeset,
but additional nodal data can be taken from the other Marvel dataset here.

The second network, \code{ison_marvel_relationships}, is a one-mode signed network
of friendships and enmities between the 53 Marvel comic book characters.
Friendships are indicated by a positive sign in the edge \code{sign} attribute,
whereas enmities are indicated by a negative sign in this edge attribute.
Additional nodal variables have been coded and included by Dr Umut Y端ksel:
\itemize{
\item \strong{Gender}: binary character, 43 "Male" and 10 "Female"
\item \strong{PowerOrigin}: binary character, 2 "Alien", 1 "Cyborg", 5 "God/Eternal",
22 "Human", 1 "Infection", 16 "Mutant", 5 "Radiation", 1 "Robot"
\item \strong{Appearances}: integer, in how many comic book issues they appeared in
\item \strong{Attractive}: binary integer, 41 1 (yes) and 12 0 (no)
\item \strong{Rich}: binary integer, 11 1 (yes) and 42 0 (no)
\item \strong{Intellect}: binary integer, 39 1 (yes) and 14 0 (no)
\item \strong{Omnilingual}: binary integer, 8 1 (yes) and 45 0 (no)
\item \strong{UnarmedCombat}: binary integer, 51 1 (yes) and 2 0 (no)
\item \strong{ArmedCombat}: binary integer, 25 1 (yes) and 28 0 (no)
}
}
\keyword{datasets}
