% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_southern_women}
\alias{ison_southern_women}
\title{Two-mode southern women (Davis, Gardner and Gardner 1941)}
\format{
\code{{igraph}} two-mode graph object with 18 women and 14 informal social
events.
}
\usage{
data(ison_southern_women)
}
\description{
Two-mode network dataset collected by Davis, Gardner and Gardner (1941)
about the attendance pattern of women at informal social events
during a 9 month period. Events and women are named.
}
\references{
Davis, A., Gardner, B., and Gardner, R. 1941. \emph{Deep South}.
Chicago: University of Chicago Press.
}
\keyword{datasets}
