% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_m182.R
\docType{data}
\name{m182}
\alias{m182}
\alias{ison_m182}
\title{Multiplex igraph of friends, social, and task ties between 16 anonymous students}
\format{
Multiplex tidygraph of friends, social, and task ties between 16 anonymous students
The edge attribute \code{friend_ties} contains friendship ties,
where \code{2} = best friends, \code{1} = friend, and \code{0} is not a friend.
\code{social_ties} consists of social interactions per hour,
and \code{task_ties} consists of task interactions per hour.
}
\source{
See also \code{data(studentnets.M182, package = "NetData")}
Larger comprehensive data set publicly available, contact Daniel A. McFarland for details.
}
\usage{
data(ison_m182)
}
\description{
M182 was an honors algebra class and friendship, social, and task ties
were collected/observed.
}
\keyword{datasets}
