% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_marvel.R
\docType{data}
\name{marvel}
\alias{marvel}
\alias{marvel_affil}
\alias{marvel_frien}
\title{Multilevel two-mode affiliation, signed one-mode networks of Marvel comicbook characters}
\format{
Two-mode igraph of 53 Marvel comicbook characters and 141 team-ups, with 683 team affiliations between them

One-mode igraph of 53 Marvel comicbook characters and 558 signed ("+" = friends, "-" = enemies) undirected ties
}
\usage{
data(marvel_affil)

data(marvel_frien)
}
\description{
Multilevel two-mode affiliation, signed one-mode networks of Marvel comicbook characters
}
\keyword{datasets}
