% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraph-package.R
\docType{package}
\name{migraph-package}
\alias{migraph}
\alias{migraph-package}
\title{migraph: Inferential Methods for Multimodal and Other Networks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of tools for testing networks. It includes functions for univariate and multivariate conditional uniform graph and quadratic assignment procedure testing, and network regression. The package is a complement to 'Multimodal Political Networks' (2021, ISBN:9781108985000), and includes various datasets used in the book. Built on the 'manynet' package, all functions operate with matrices, edge lists, and 'igraph', 'network', and 'tidygraph' objects, and on one-mode and two-mode (bipartite) networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stocnet.github.io/migraph/}
  \item Report bugs at \url{https://github.com/stocnet/migraph/issues}
}

}
\author{
\strong{Maintainer}: James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]

Authors:
\itemize{
  \item James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]
}

Other contributors:
\itemize{
  \item Henrique Sposito (\href{https://orcid.org/0000-0003-3420-6085}{ORCID}) (IHEID) [contributor]
  \item Jael Tan \email{jael.tan@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-6234-9764}{ORCID}) (IHEID) [contributor]
  \item Bernhard Bieri \email{bernhard.bieri@graduateinstitute.ch} (\href{https://orcid.org/0000-0001-5943-9059}{ORCID}) [contributor]
}

}
\keyword{internal}
