% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mpn.R
\docType{data}
\name{mpn_cow_separate}
\alias{mpn_cow_separate}
\alias{mpn_cow_trade}
\alias{mpn_cow_igo}
\title{One-mode interstate trade relations and two-mode state membership in IGOs (COW)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> -- # Interstate trade network --------------------------------------------------
#> # A labelled, weighted, directed network of 116 states and 10730 trade arcs
#> 
#> -- Nodes
#> # A tibble: 116 x 1
#>   label                   
#>   <chr>                   
#> 1 United States of America
#> 2 Canada                  
#> 3 Cuba                    
#> 4 Dominican Republic      
#> 5 Jamaica                 
#> 6 Trinidad and Tobago     
#> # i 110 more rows
#> 
#> -- Missings
#> # A tibble: 759 x 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     1    31     NA
#> 2     2    31     NA
#> 3     3    31     NA
#> 4     3    41     NA
#> 5     3    59     NA
#> 6     3    66     NA
#> # i 753 more rows
#> 
#> -- Ties
#> # A tibble: 10,730 x 3
#>    from    to  weight
#>   <int> <int>   <dbl>
#> 1     1     2 180387 
#> 2     1     3    587.
#> 3     1     4   5511.
#> 4     1     5   1896.
#> 5     1     6   2188.
#> 6     1     7 123677 
#> # i 10,724 more rows
#> 
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{#> -- # State IGO membership network ----------------------------------------------
#> # A labelled, two-mode network of 112 states and 40 IGOs and 839 membership
#> ties
#> 
#> -- Nodes
#> # A tibble: 152 x 4
#>   label       mode   polity2 scope
#>   <chr>       <chr>    <dbl> <chr>
#> 1 Afghanistan states      -7 <NA> 
#> 2 Albania     states       5 <NA> 
#> 3 Algeria     states      -3 <NA> 
#> 4 Angola      states      -6 <NA> 
#> 5 Argentina   states       8 <NA> 
#> 6 Australia   states      10 <NA> 
#> # i 146 more rows
#> 
#> -- Ties
#> # A tibble: 839 x 2
#>    from    to
#>   <int> <int>
#> 1     1   113
#> 2     1   114
#> 3     1   115
#> 4     1   116
#> 5     1   117
#> 6     1   118
#> # i 833 more rows
#> 
}\if{html}{\out{</div>}}
}
\source{
The Correlates of War Project. 2012. \emph{Trade}.

Barbieri, Katherine and Omar Keshk. 2012.
Correlates of War Project Trade Data Set Codebook, Version 3.0.

The Correlates of War Project. 2019. \emph{Intergovernmental Organization v3}.
}
\usage{
data(mpn_cow_trade)

data(mpn_cow_igo)
}
\description{
\link{mpn_cow} combines these two datasets into one multilevel network.
They are kept alongside it because \code{{manynet}} currently reads a network with
two named modes as undirected, which symmetrises the trade layer of \code{mpn_cow},
and because \code{to_layer()} does not recover that layer either.
See stocnet/manynet#170 and stocnet/manynet#171.
}
\details{
\code{mpn_cow_trade} is a one-mode matrix representing the trade relations between 116 states.
The data is derived from the Correlates of War Project (COW) Trade Dataset (v3.0),
which contains the annual dyadic and national trade figures for states (listed in COW)
between 1870 to 2009.
This network is based only on the dyadic trade figures in 2009 for the 116 states
listed in Appendix 7.1 in
\href{https://www.cambridge.org/core/books/multimodal-political-networks/43EE8C192A1B0DCD65B4D9B9A7842128}{\emph{Multimodal Political Networks}}
(Knoke et al., 2021).
The value in each cell of the matrix represents the value of exports
from the 116 row states to the 116 column states.

\code{mpn_cow_igo} is a two-mode graph representing the membership of 116 states
in 40 intergovernmental organizations (IGOs).
The data is derived from the Correlates of War Project (COW)
Intergovernmental Organizations Dataset (v3.0),
which contains information about intergovernmental organizations from 1815-2014,
such as founding year and membership.
This network contains only a subset of the states and IGOs listed in COW,
with 116 states listed in Appendix 7.1 in \href{https://www.cambridge.org/core/books/multimodal-political-networks/43EE8C192A1B0DCD65B4D9B9A7842128}{\emph{Multimodal Political Networks}}
and 40 IGOs from Table 7.1 in \href{https://www.cambridge.org/core/books/multimodal-political-networks/43EE8C192A1B0DCD65B4D9B9A7842128}{\emph{Multimodal Political Networks}}
that also overlap with the COW dataset (Knoke et al., 2021).
}
\references{
Barbieri, Katherine, Omar M. G. Keshk, and Brian Pollins. 2009.
“TRADING DATA: Evaluating our Assumptions and Coding Rules.”
\emph{Conflict Management and Peace Science} 26(5): 471-491.
\doi{10.1177/0738894209343887}.

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021.
\href{https://www.cambridge.org/core/books/multimodal-political-networks/43EE8C192A1B0DCD65B4D9B9A7842128}{\emph{Multimodal Political Networks}}.
Cambridge University Press. Cambridge University Press.

Pevehouse, Jon C.W., Timothy Nordstron, Roseanne W McManus, Anne Spencer Jamison. 2020.
“Tracking Organizations in the World: The Correlates of War IGO Version 3.0 datasets”.
\emph{Journal of Peace Research} 57(3): 492-503.
\doi{10.1177/0022343319881175}.
}
\keyword{datasets}
