% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mpn.R
\docType{data}
\name{mpn_ryanair}
\alias{mpn_ryanair}
\title{One-mode EU policy influence network, June 2004 (Christopoulos 2006)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> -- # EU policy influence network -----------------------------------------------
#> # A labelled, directed network of 20 policy actors and 177 interaction arcs
#> 
#> -- Nodes
#> # A tibble: 20 x 1
#>   label         
#>   <chr>         
#> 1 1 AER         
#> 2 2 AER         
#> 3 5 AER/COR     
#> 4 7 RYANAIR     
#> 5 8 DG TRANSPORT
#> 6 9 COR         
#> # i 14 more rows
#> 
#> -- Ties
#> # A tibble: 177 x 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     1     3
#> 3     1     4
#> 4     1     5
#> 5     1     6
#> 6     1     7
#> # i 171 more rows
#> 
}\if{html}{\out{</div>}}
}
\source{
Christopoulos, Dimitrios C. 2006.
“Relational Attributes of Political Entrepreneurs: a Network Perspective.”
\emph{Journal of European Public Policy} 13(5): 757–78.
\doi{10.1080/13501760600808964}.

Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021.
\href{https://www.cambridge.org/core/books/multimodal-political-networks/43EE8C192A1B0DCD65B4D9B9A7842128}{\emph{Multimodal Political Networks}}.
Cambridge University Press. Cambridge University Press.
}
\usage{
data(mpn_ryanair)
}
\description{
Network of anonymised actors reacting to the Ryanair/Charleroi decision of
the EU Commission in February 2004.
The relationships mapped comprise an account of public records of interaction
supplemented with the cognitive network of key informants.
Examination of relevant communiques, public statements and a number of
off-the-record interviews provides confidence
that the network mapped closely approximated interactions between 29 January
and 12 February 2004. The time point mapped is at the height of influence and
interest intermediation played by actors in the AER,
a comparatively obscure body representing the interests of a number of
European regional bodies at the EU institutions.
}
\keyword{datasets}
