% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_to.R
\name{na}
\alias{na}
\alias{na_to_null}
\alias{na_to_mean}
\title{Tools for imputing missing tie data}
\usage{
na_to_null(object)

na_to_mean(object)
}
\description{
These functions offer tools for imputing missing tie data.
}
\section{Functions}{
\itemize{
\item \code{na_to_null()}: Impute missing tie data as null,
the modal value in sparse social networks.

\item \code{na_to_mean()}: Impute missing tie data as
the mean value in the network.

}}
\examples{
missTest <- ison_adolescents \%>\% 
   add_tie_attribute("weight", c(1,NA,NA,1,1,1,NA,NA,1,1)) \%>\% 
   as_matrix
missTest
na_to_null(missTest)
na_to_mean(missTest)
}
\seealso{
Other manipulations: 
\code{\link{add}},
\code{\link{as}()},
\code{\link{grab}},
\code{\link{reformat}},
\code{\link{split}()},
\code{\link{transform}()}
}
\concept{manipulations}
