% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_layout.R
\name{network_map}
\alias{network_map}
\title{Plot geographical networks}
\usage{
network_map(object, date, theme = "light")
}
\arguments{
\item{object}{Unimodal geographical network. Needs to contain ISO3c country
IDs.}

\item{date}{String date at which the network snapshot was taken e.g.
"2010-01-01". Used by \code{{cshapes}} to plot the correct map hence needs to
be between 1886 and 2021.}

\item{theme}{Theme you would like to use to plot the graph. Available themes
are "light", "dark", and "earth".}
}
\value{
A map of a country level geographical network.
}
\description{
Creates a plot of the a unimodal geographical network.
}
\examples{
# Load the unimodal network of environmental agreements signed in 2010
# Made from {manyenviron} using data from ECOLEX.
membership <- migraph:::membership

# Plot the network at the specified date
# Light theme
network_map(membership, date = "2010-01-01", theme = "light") +
  labs(title = "International Environmental Treaties 2010",
       subtitle = "Ecolex data",
       caption = "Created with love by {migraph}")
# Dark theme
network_map(membership, date = "2010-01-01", theme = "dark") +
  labs(title = "International Environmental Treaties 2010",
       subtitle = "Ecolex data",
       caption = "Created with love by {migraph}")
# Earth theme
network_map(membership, date = "2010-01-01", theme = "earth") +
  labs(title = "International Environmental Treaties 2010",
       subtitle = "Ecolex data",
       caption = "Created with love by {migraph}")
}
