% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.R
\name{node_components}
\alias{node_components}
\title{Identifying nodes' component membership}
\usage{
node_components(object, method = c("weak", "strong"))
}
\arguments{
\item{object}{a migraph-consistent object}

\item{method}{For directed networks,
either \code{weak} if edge direction is irrelevant,
or \code{strong} if edge direction is salient.
Ignored if network undirected.}
}
\description{
Identifying nodes' component membership
}
