% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_core.R
\name{node_core}
\alias{node_core}
\title{Core-periphery clustering algorithms}
\usage{
node_core(object)
}
\description{
Core-periphery clustering algorithms
}
\examples{
mpn_elite_usa_advice \%>\% as_tidygraph \%>\% 
 mutate(corep = node_core(mpn_elite_usa_advice)) \%>\% 
 autographr(node_color = "corep")
}
\references{
Lip, Sean Z. W. 2011.
“A Fast Algorithm for the Discrete Core/Periphery Bipartitioning Problem.”
}
