% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_correlation.R
\name{node_correlation}
\alias{node_correlation}
\alias{to_correlation}
\title{Node correlation}
\usage{
to_correlation(.data)
}
\arguments{
\item{.data}{An object of a \code{{manynet}}-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}
}
\description{
This function performs a Pearson pairwise correlation on a given matrix or network data.
It includes a switch:
whereas for a two-mode network it will perform a regular correlation,
including all rows,
for an undirected network it will perform a correlation on a matrix
with the diagonals removed,
for a reciprocated network it will include the difference
between reciprocated ties,
and for complex networks it will include also the difference
between the self ties in each pairwise calculation.
This function runs in \eqn{O(mn^2)} complexity.
}
\seealso{
Other motifs: 
\code{\link{brokerage_census}},
\code{\link{network_census}},
\code{\link{node_census}}
}
\concept{motifs}
