% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_diffusion.R
\name{node_diffusion}
\alias{node_diffusion}
\alias{node_adoption_time}
\alias{node_adopter}
\alias{node_thresholds}
\alias{node_infection_length}
\alias{node_exposure}
\alias{node_is_exposed}
\title{Diffusion metrics for nodes}
\usage{
node_adoption_time(diff_model)

node_adopter(diff_model)

node_thresholds(diff_model)

node_infection_length(diff_model)

node_exposure(.data, mark)

node_is_exposed(.data, mark)
}
\arguments{
\item{diff_model}{A valid network diffusion model,
as created by \code{as_diffusion()} or \code{play_diffusion()}.}
}
\description{
These functions allow measurement of various features of
a diffusion process:
\itemize{
\item \code{node_adoption_time()}: Measures nodes' time of adoption/infection
\item \code{node_adopter()}: Classifies membership of nodes into diffusion categories
\item \code{node_thresholds()}: Measures nodes' thresholds from the amount
of exposure they had when they became infected
\item \code{node_infection_length()}: Measures the average length nodes that become
infected remain infected in a compartmental model with recovery
\item \code{node_exposure()}: Measures how many exposures nodes have to
a given mark
\item \code{node_is_exposed()}: Marks the nodes that are susceptible,
i.e. are in the immediate neighbourhood of given mark vector
}
}
\examples{
  smeg <- manynet::generate_smallworld(15, 0.025)
  smeg_diff <- play_diffusion(smeg, recovery = 0.2)
  plot(smeg_diff)
  autographs(smeg_diff)
  # autographd(smeg_diff)
  (adopts <- node_adopter(smeg_diff))
  summary(adopts)
  summary(node_adoption_time(smeg_diff), membership = adopts)
  summary(node_thresholds(smeg_diff), membership = adopts)
  summary(node_infection_length(smeg_diff))
  network_infection_length(smeg_diff)
  network_transmissibility(smeg_diff)
  network_reproduction(smeg_diff)
# example code
  smeg <- manynet::generate_smallworld(15, 0.025)
  smeg_diff <- play_diffusion(smeg, recovery = 0.2)
  (adopts <- node_adopter(smeg_diff))
  summary(adopts)
}
\references{
Valente, Tom W. (1995). \emph{Network models of the diffusion of innovations}
(2nd ed.). Cresskill N.J.: Hampton Press.
}
\seealso{
Other measures: 
\code{\link{between_centrality}},
\code{\link{close_centrality}},
\code{\link{closure}},
\code{\link{cohesion}()},
\code{\link{degree_centrality}},
\code{\link{eigenv_centrality}},
\code{\link{features}},
\code{\link{heterogeneity}},
\code{\link{hierarchy}},
\code{\link{holes}},
\code{\link{net_diffusion}}
}
\concept{measures}
