% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{node_triad_census}
\alias{node_triad_census}
\title{Triad census by nodes}
\usage{
node_triad_census(object)
}
\arguments{
\item{object}{a migraph-consistent object}
}
\description{
Triad census by nodes
}
\examples{
node_triad_census(adolescent_society)
}
