% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_community.R
\name{partition_kernighanlin}
\alias{partition_kernighanlin}
\title{Kernighan-Lin algorithm for graph partitioning}
\usage{
partition_kernighanlin(object)
}
\arguments{
\item{object}{}
}
\description{
This function partitions a graph into two communities.
}
\examples{
partition_kernighanlin(to_unsigned(ison_marvel_relationships, keep = "positive"))
partition_kernighanlin(ison_adolescents)
}
