% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blockmodel}
\alias{plot.blockmodel}
\title{ggplot2-based plotting of blockmodel results}
\usage{
\method{plot}{blockmodel}(x, ...)
}
\arguments{
\item{x}{A blockmodel-class object.}

\item{...}{Additional arguments passed on to ggplot2.}
}
\description{
ggplot2-based plotting of blockmodel results
}
\examples{
usa_concor <- blockmodel_concor(mpn_elite_usa_advice)
plot(usa_concor)
}
