% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.igraph}
\alias{plot.igraph}
\title{Plotting of one-mode and two-mode graphs}
\usage{
\method{plot}{igraph}(x, ...)
}
\arguments{
\item{x}{A migraph-compatible object, especially an igraph graph object.}

\item{...}{Additional arguments passed on to igraph.}
}
\description{
Plotting of one-mode and two-mode graphs
}
\examples{
mat1 <- create_ring(5,10)
plot(mat1)
}
\concept{plotting}
