% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_2x2}
\alias{plot_2x2}
\title{Plotting two-by-two plots}
\usage{
plot_2x2(dat)
}
\arguments{
\item{dat}{A data.frame}
}
\value{
A plot
}
\description{
Better default plotting of values traced through a two-by-two than ggplot2
}
\details{
Makes for clearer communication of structure than ggplot2 defaults
}
\examples{
\dontrun{
plot_2x2(dat)
}
}
