% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_globalnet}
\alias{plot_globalnet}
\title{Plotting multilevel networks on a globe}
\usage{
plot_globalnet(edges, actors, time)
}
\arguments{
\item{edges}{A data frame that includes (e.g. country) actors in the first column,
and institutional ties in the second column.}

\item{actors}{A data frame that includes 'StatID' variable (for identification),
and 'Latitude' and 'Longitude' variables for plotting.}

\item{time}{A date to render network for.}
}
\value{
An html/js viewer pane
}
\description{
Plotting of a two-mode network with multilevel structure revealed
}
\details{
Will separate and plot bilateral and pluri-/mini-/multilateral agreements
with different colours. It will also plot a pin for the centroid marker of each
multilateral agreement.
}
\examples{
\dontrun{
plot_globalnet(mat)
}
}
