% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_multilevel}
\alias{plot_multilevel}
\title{Plotting multilevel networks}
\usage{
plot_multilevel(mat, levels = NULL)
}
\arguments{
\item{mat}{An incidence matrix}
}
\value{
A plot
}
\description{
Plotting of a two-mode network with multilevel structure revealed
}
\details{
Removes all isolates to clarify structure. 
Relies on Fruchterman-Rheingold to give initial coordinates, 
which is then rotated around the x-axis.
TODO: Allow more than two levels to be plotted at once.
}
\examples{
\dontrun{
plot_multilevel(create_random(10,12,0.2))
}
}
