% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_releases.R
\name{plot_releases}
\alias{plot_releases}
\title{A plotting function that visualises historical milestones/releases}
\source{
https://benalexkeen.com/creating-a-timeline-graphic-using-r-and-ggplot2/
}
\usage{
plot_releases(repo)
}
\arguments{
\item{repo}{the github repository to track, e.g. "snlab-ch/migraph"}
}
\value{
A ggplot graph object
}
\description{
The function will take a data frame that details this information,
or more usefully, a Github repository listing.
}
\details{
The function creates a project timeline graphic using ggplot2
with historical milestones and milestone statuses gathered from a
specified GitHub repository.
}
\examples{
if(!httr::http_error("https://api.github.com/repos/snlab-ch/migraph/releases")){
plot_releases("snlab-ch/migraph")
}
}
