% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_twomode}
\alias{plot_twomode}
\title{Plotting two-mode networks}
\usage{
plot_twomode(mat, attr = NULL, greyscale = T, ...)
}
\arguments{
\item{mat}{A matrix}

\item{greyscale}{Whether to use black/white or blue/green, Default: T}

\item{...}{Any other arguments passed to igraph.plotting}
}
\value{
A plot
}
\description{
Better default plotting of two-mode networks than igraph
}
\details{
Makes for clearer communication of structure than igraph defaults
for two-mode networks by providing default colours, shapes, size, and frames,
and avoiding labels.
}
\examples{
\dontrun{
plot_twomode(mat)
}
}
