% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_conversion.R
\name{project_list}
\alias{project_list}
\title{if(is.list(edgelists)) edgelists <- dplyr::bind_rows(edgelists)
  
  do.call(lapply(edgelists, function(x) {
    table(x[,1], x[,2])
  }), rbind)
  
  
  do.call(plyr::rbind.fill, test)
  
  data.table::rbindlist(lapply(edgelists, function(x) {
    as.data.frame(table(x[,1], x[,2]))
  }))
  
  test <- plyr::rbind.fill(lapply(edgelists, function(x) {
    as.data.frame(table(x[,1], x[,2]), stringsAsFactors = F)
  }))
  
  mat <- matrix(0,length(unique(c(test[,1], test[,2]))), length(unique(c(test[,1], test[,2]))))
  rownames(mat) <- colnames(mat) <- unique(c(test[,1], test[,2]))
  mat[match(test[test[,3]>0,1],rownames(mat)), match(test[test[,3]>0,2],colnames(mat))] <- 1
  
  mat[match(test[,1],rownames(mat)), match(test[,2],colnames(mat))] <- test[,3]
  mat[1,11]
  
  
  
  
  data <- list(data.frame(numbers=1:10,
                          letters=letters[1:10]),
               data.frame(letters=letters[1:10],
                          colours=rainbow(10)))
  
  Reduce(rbind, test)
  
  as.matrix(Matrix::bdiag(test))
  
  test[[1]] %+% test[[2]]
  
Transform a two-mode event list into a one-mode event list}
\usage{
project_list(eventlist)
}
\arguments{
\item{eventlist}{a two-mode event list}
}
\value{
a one-mode event list
}
\description{
This function relies on `tidyverse` functions to quickly and efficiently
project a two-mode event list into a one-mode event list of nodes from 
the first mode that share connections to nodes in the second mode.
}
