% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{read_graph}
\alias{read_ucinet}
\alias{write_ucinet}
\title{Reading from/writing to external formats}
\usage{
read_graph(
  file = file.choose(),
  csv = c(",", ";"),
  to = c("igraph", "tidygraph", "network"),
  ...
)

read_ucinet(header_file)

write_ucinet(
  object,
  filename = deparse(substitute(object)),
  name = deparse(substitute(object))
)
}
\arguments{
\item{file, header_file}{A character string giving the path to the header
(.##h) file. If the function is called without a header_file specified,
an OS-specific file picker is opened to help users select it.}

\item{csv}{Allows users to specify whether their csv file is comma or column
separated.}

\item{to}{Selects the output formal the graph is converted to when imported.}

\item{...}{Additional parameters passed to the read/write function.}

\item{object}{A migraph-consistent object to be exported.}

\item{filename}{UCINET filename (without ## extension).
By default the files will have the same name as the object
and be saved to the working directory.}

\item{name}{name of matrix to be known in UCINET.
By default the name will be the same as the object.}

\item{graph}{Graph object to be written from}

\item{path}{Specify the path where the file will be written.}
}
\value{
By default, \code{\link[=read_ucinet]{read_ucinet()}} will import into an igraph format.
\code{\link[=read_graph]{read_graph()}} allows users to choose the format among those supported
by \code{{migraph}}. Note that graphs can be easily coerced into other formats
with \code{{migraph}}'s \code{as_} methods.

A pair of UCINET files in V6404 file format (.##h, .##h)
}
\description{
These functions read and write graphs from and to a variety of formats.
The \code{read_ucinet()} and \code{write_ucinet()} functions import from and export
to UCINET network files.
}
\details{
These functions only work with relatively recent UCINET
file formats, e.g. type 6406 files.
To import earlier UCINET file types, you will need to update them first.

To import multiple matrices packed into a single UCINET file,
you will need to unpack them and convert them one by one.
}
\examples{
\dontrun{
# import Roethlisberger & Dickson's horseplay game data set:
horseplay <- read_ucinet("WIRING-RDGAM.##h")
}
\dontrun{
# export it again to UCINET under a different name:
write.ucinet(horseplay,"R&D-horseplay")
}
}
\seealso{
\link{convert}
}
\author{
Christian Steglich, 18 June 2015
}
