% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_to.R
\name{split}
\alias{split}
\alias{to_egos}
\alias{to_subgraphs}
\alias{to_components}
\alias{to_waves}
\alias{to_slices}
\alias{from_waves}
\title{Tools for splitting and joining networks, graphs, and matrices}
\usage{
to_egos(object, max_dist = 1, min_dist = 0)

to_subgraphs(object, attribute)

to_components(object)

to_waves(.data, attribute = "wave", delete.vertices = FALSE)

to_slices(.data, attribute, slice, delete.vertices = FALSE)

from_waves(.data)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}

\item{max_dist}{The maximum breadth of the neighbourhood.
By default 1.}

\item{min_dist}{The minimum breadth of the neighbourhood.
By default 0.
Increasing this to 1 excludes the ego,
and 2 excludes ego's direct alters.}

\item{attribute}{Character string indicating attribute used to slice data.}

\item{delete.vertices}{Would you like to remove vertices that do not have
any adjacent edges for each slice?}

\item{slice}{Character string or list indicating the date(s) or integer(s)
from attribute used to slice data into subgraphs.}
}
\description{
These functions offer tools for splitting migraph-consistent objects
(matrices, igraph, tidygraph, or network objects).
Splitting means that the returned object will be a list of objects.
Joining expects a list of objects and returns a network object.

#' @export
to_joined.list <- function(.data,
from = c("waves")){
from <- match.arg(from)
ann <- switch(from,
"waves" = lapply(seq_along(.data), function(x) mutate_ties(.data[\link{x}], wave = x)))
out <- ann[\link{1}]
for (i in seq_along(ann)\link{-1}){
out <- join_ties(out, ann[\link{i}], "wave")
}
out
}
}
\section{Functions}{
\itemize{
\item \code{to_egos()}: Returns a list of ego (or focal)
networks.

\item \code{to_subgraphs()}: Returns a list of subgraphs
on some given attribute.

\item \code{to_components()}: Returns a list of the components
in a network.

\item \code{to_waves()}: Returns a network
with some discrete observations over time
into a list of those observations.

\item \code{to_slices()}: Returns a list of a network
with some continuous time variable
at some time slice(s).

\item \code{from_waves()}: Returns a single network object
from a list of waves.

}}
\examples{
autographs(to_egos(ison_adolescents))
autographs(to_egos(ison_adolescents,2))
to_components(ison_marvel_relationships)
ison_adolescents \%>\%
  activate(edges) \%>\%
  mutate(wave = sample(1:4, 10, replace = TRUE)) \%>\%
  to_waves(attribute = "wave")
ison_adolescents \%>\%
  activate(edges) \%>\%
  mutate(beg = sample(1:4, 10, replace = TRUE),
  end = sample(5:8, 10, replace = TRUE)) \%>\%
  to_waves(attribute = c("beg", "end"))
ison_adolescents \%>\%
  activate(edges) \%>\%
  mutate(wave = sample(1:6, 10, replace = TRUE)) \%>\%
  to_slices(attribute = "wave", slice = c(2, 4, 6))
ison_adolescents \%>\%
  activate(edges) \%>\%
  mutate(wave = sample(1:4, 10, replace = TRUE)) \%>\%
  to_waves(attribute = "wave") \%>\%
  from_waves()
}
\seealso{
Other manipulations: 
\code{\link{add}},
\code{\link{as}()},
\code{\link{grab}},
\code{\link{na}},
\code{\link{reformat}},
\code{\link{transform}()}
}
\concept{manipulations}
