% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_motif.R
\name{summary.node_motif}
\alias{summary.node_motif}
\title{#' @export
print.node_motif <- function(x, ...,
max.length = 6,
digits = 3){
if(any(attr(x, "mode"))){
y <- x\link{attr(x, "mode")}
y <- y\link{max(1, ((length(y)-max.length)+1)):length(y)}
z <- x\link{!attr(x, "mode")}
z <- z\link{1:min(length(z), max.length)}
class(z) <- "numeric"
z <- format(z, digits = digits)
class(y) <- "numeric"
y <- format(y, digits = digits)
print(noquote(format(c(z,
paste("+", length(x) - (length(z) + length(y)),
"others"), y))))
} else {
z <- x\link{1:min(length(x), max.length)}
class(z) <- "numeric"
z <- format(z, digits = digits)
print(noquote(format(c(z,
paste("+", length(x) - length(z),
"others")))))
}
}}
\usage{
\method{summary}{node_motif}(object, ..., membership, FUN = mean)
}
\description{
#' @export
print.node_motif <- function(x, ...,
max.length = 6,
digits = 3){
if(any(attr(x, "mode"))){
y <- x\link{attr(x, "mode")}
y <- y\link{max(1, ((length(y)-max.length)+1)):length(y)}
z <- x\link{!attr(x, "mode")}
z <- z\link{1:min(length(z), max.length)}
class(z) <- "numeric"
z <- format(z, digits = digits)
class(y) <- "numeric"
y <- format(y, digits = digits)
print(noquote(format(c(z,
paste("+", length(x) - (length(z) + length(y)),
"others"), y))))
} else {
z <- x\link{1:min(length(x), max.length)}
class(z) <- "numeric"
z <- format(z, digits = digits)
print(noquote(format(c(z,
paste("+", length(x) - length(z),
"others")))))
}
}
}
