% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_test.R
\name{test_tutorials}
\alias{test_tutorials}
\title{Test tutorials}
\usage{
test_tutorials(path, quiet = TRUE)
}
\arguments{
\item{path}{Character vector of the paths to the tutorials to be knitted.}

\item{quiet}{Logical, whether to suppress messages from \code{render()}.}
}
\value{
No return value, called for side effects.
}
\description{
For our purposes, "testing" a tutorial means being able to
(successfully) run \code{render()} on it.
This function renders the tutorial provided in \code{path}.
There is no check to see if the rendered file looks OK.
If a tutorial fails to render, then an error will be generated which will
propagate to the caller.
}
\author{
David Kane, see tutorial.helpers
}
