% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{threemode_clustering}
\alias{threemode_clustering}
\title{Three-mode clustering}
\usage{
threemode_clustering(mat1, mat2)
}
\arguments{
\item{mat1}{The first two-mode matrix}

\item{mat2}{The second two-mode matrix}
}
\description{
This function allows you to calculate how much three-mode clustering there is.
}
\examples{
threemode_clustering(matrix)
}
\concept{three-mode functions}
