% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to}
\alias{to}
\alias{to_unweighted}
\alias{to_undirected}
\alias{to_onemode}
\alias{to_main_component}
\alias{to_uniplex}
\alias{to_unsigned}
\alias{to_simplex}
\alias{to_named}
\alias{to_unnamed}
\alias{to_multilevel}
\alias{to_mode1}
\alias{to_mode2}
\alias{to_edges}
\title{Tools for reformatting networks, graphs, and matrices}
\usage{
to_unweighted(object, threshold = 1)

to_undirected(object)

to_onemode(object)

to_main_component(object)

to_uniplex(object, edge)

to_unsigned(object, keep = c("positive", "negative"))

to_simplex(object)

to_named(object, names = NULL)

to_unnamed(object)

to_multilevel(object)

to_mode1(object)

to_mode2(object)

to_edges(object)
}
\arguments{
\item{object}{A matrix, \code{{igraph}} graph, \code{{tidygraph}} tbl_graph, or
\code{{network}} object.}

\item{threshold}{For a matrix, the threshold to binarise/dichotomise at.}

\item{edge}{the name of an edge attribute to retain from a graph}

\item{keep}{in the case of a signed network, whether to retain
the "positive" or "negative" ties}

\item{names}{Character vector of the node names. Null by default.}
}
\value{
All \code{to_} functions return an object of the same class as that provided.
So passing it an igraph object will return an igraph object
and passing it a network object will return a network object,
with certain modifications as outlined for each function.
\itemize{
\item \code{to_undirected()} returns an object that has any edge direction removed
\item \code{to_onemode()} returns an object that has any type/mode attributes removed,
but otherwise includes all the same nodes and ties.
Note that this is not the same as \code{to_mode1()} or \code{to_mode2()},
which return only some of the nodes and new ties established by coincidence.
\item \code{to_simplex()} returns an object that has all loops or self-ties removed
}
}
\description{
These functions offer tools for transforming certain properties
of migraph-consistent objects
(that is, matrices, igraph, tidygraph, or network objects).
Unlike the \link{as} group of functions,
these functions always return the same object type as they are given,
and are concerned with transforming these object's properties.
}
\details{
Since some modifications are easier to implement for some objects than others,
here are the currently implemented modifications:\tabular{lccccc}{
   to_ \tab edgelists \tab matrices \tab igraph \tab tidygraph \tab network \cr
   unweighted \tab X \tab X \tab X \tab X \tab X \cr
   undirected \tab  \tab X \tab X \tab X \tab X \cr
   unsigned \tab  \tab  \tab X \tab X \tab  \cr
   uniplex \tab  \tab  \tab X \tab X \tab  \cr
   unnamed \tab X \tab X \tab X \tab X \tab X \cr
   named \tab  \tab X \tab X \tab X \tab X \cr
   simplex \tab  \tab X \tab X \tab X \tab  \cr
   main_component \tab  \tab  \tab X \tab X \tab X \cr
   onemode \tab  \tab  \tab X \tab X \tab  \cr
   multilevel \tab  \tab X \tab X \tab X \tab  \cr
}
}
\section{Functions}{
\itemize{
\item \code{to_unweighted}: Returns an object that has all edge weights removed

\item \code{to_main_component}: Returns an object that includes only the main component
without any smaller components or isolates

\item \code{to_uniplex}: Returns an object that includes only a single type of tie

\item \code{to_unsigned}: Returns a network with either just the "positive" ties
or just the "negative" ties

\item \code{to_named}: Returns an object that has random vertex names added

\item \code{to_unnamed}: Returns an object with all vertex names removed

\item \code{to_mode1}: Results in a weighted one-mode object
that retains the column nodes from a two-mode object,
and weights the ties between them on the basis of
their joint ties to nodes in the first mode (rows).

\item \code{to_mode2}: Results in a weighted one-mode object
that retains the row nodes from a two-mode object,
and weights the ties between them on the basis of
their joint ties to nodes in the second mode (columns)

\item \code{to_edges}: Returns a matrix (named if possible)
where the edges are the nodes
}}

\examples{
  autographr(to_mode2(ison_southern_women)) +
  autographr(to_unweighted(to_mode2(ison_southern_women)))
to_undirected(ison_adolescents)
to_onemode(ison_marvel_teams)
autographr(ison_marvel_relationships) +
  autographr(to_main_component(ison_marvel_relationships))
autographr(ison_algebra_class) + 
autographr(to_uniplex(ison_algebra_class, "friend_tie"))
(autographr(ison_marvel_relationships) |
autographr(to_main_component(ison_marvel_relationships))) /
  (autographr(to_main_component(to_unsigned(ison_marvel_relationships, "positive"))) |
  autographr(to_main_component(to_unsigned(ison_marvel_relationships, "negative"))))
to_simplex(ison_algebra_class)
autographr(ison_algebra_class) + ggtitle("Original") + 
  autographr(to_named(ison_algebra_class)) + ggtitle("Named") + 
  autographr(to_unnamed(to_named(ison_algebra_class))) + ggtitle("Unnamed")
to_multilevel(mpn_elite_usa_advice)
autographr(ison_southern_women) /
 (autographr(to_mode1(ison_southern_women)) |
 autographr(to_mode2(ison_southern_women)))
autographr(ison_adolescents) +
  autographr(to_edges(ison_adolescents))
}
\seealso{
Other manipulation: 
\code{\link{coercion}},
\code{\link{is}()}
}
\concept{manipulation}
