% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{triad_census}
\alias{triad_census}
\alias{node_triad_census}
\alias{cluster_triad_census}
\title{Triad census by nodes or clusters}
\usage{
node_triad_census(object)

cluster_triad_census(object, clusters)
}
\arguments{
\item{object}{a migraph-consistent object}

\item{clusters}{a vector of cluster assignment}
}
\description{
Triad census by nodes or clusters
}
\examples{
node_triad_census(ison_coleman)
}
