% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_run.R
\name{tutorials}
\alias{tutorials}
\alias{run_tute}
\alias{extract_tute}
\title{Open and extract code from tutorials}
\usage{
run_tute(tute)

extract_tute(tute)
}
\arguments{
\item{tute}{String, name of the tutorial (e.g. "tutorial2").}
}
\description{
These functions make it easy to use the tutorials
in the \code{{manynet}} and \code{{migraph}} packages:
\itemize{
\item \code{run_tute()} runs a \code{{learnr}} tutorial from
either the \code{{manynet}} or \code{{migraph}} packages,
wraps \code{learnr::run_tutorial()} with some convenience.
\item \code{extract_tute()} extracts and opens just the solution code
from a \code{{manynet}} or \code{{migraph}} tutorial,
saving the .R script to the current working directory.
}
}
