% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallworld.R
\name{twomode.clustering}
\alias{twomode.clustering}
\title{Two-mode clustering}
\usage{
twomode.clustering(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
This function allows you to calculate how much two-mode clustering there is.
}
\examples{
twomode.clusterings(matrix)
}
\keyword{two-mode}
