% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallworld.R
\name{twomode.lattice}
\alias{twomode.lattice}
\title{Two-mode lattice}
\usage{
twomode.lattice(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
This function allows you to express your love of lattices.
}
\examples{
twomode.lattice(matrix)
}
\keyword{two-mode}
