% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_2x2}
\alias{twomode_2x2}
\title{Two-mode two-by-two analysis}
\usage{
twomode_2x2(node1, node2, ties, attr1, start, end)
}
\arguments{
\item{node1}{First nodeset object (will become matrix rows)}

\item{node2}{Second nodeset object (will become matrix columns)}

\item{ties}{Edgelist of affiliations/memberships of first nodeset in second nodeset}

\item{attr1}{Object containing information on an attribute by year, in long format}

\item{start}{Integer, for example a year like 1960}

\item{end}{Integer, for example a year like 2010}
}
\value{
A data frame with two columns (Coherence and Dominance)
}
\description{
This function calculates coherence and dominance for two-mode networks (and attributes)
extracted for each year between the given dates
}
\details{
Note that this function has defaults that make sense for use with 
the gnevar datasets. Please contact me if you would like to make this function
more general.
}
\examples{
\dontrun{
library(gnevar)
library(wbstats)
mil_data <- wb(country = unique(stat_actor$StatID), indicator = "MS.MIL.XPND.CN", startdate = 1960, enddate = 2018)
ally_topo <- twomode_2x2(stat_actor, ally_agree, ally_membs, mil_data, 1960, 2018)
plot_2x2(ally_topo)
}
}
\seealso{
twomode_clustering

twomode_dominance

Other two-mode functions: \code{\link{twomode_clustering}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_dominance}},
  \code{\link{twomode_fragmentation}},
  \code{\link{twomode_lattice}},
  \code{\link{twomode_smallworld}}
}
