% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_centralization_between}
\alias{twomode_centralization_between}
\title{Two-mode betweenness centralization}
\usage{
twomode_centralization_between(graph)
}
\arguments{
\item{graph}{An igraph graph}
}
\description{
This function allows you to calculate how (betweenness) centralized a two-mode graph is.
}
\examples{
twomode_centralization_between(graph)
}
\seealso{
Other two-mode: \code{\link{twomode_clustering}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_dominance}},
  \code{\link{twomode_fragmentation}},
  \code{\link{twomode_lattice}}
}
\keyword{two-mode}
