% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_components}
\alias{twomode_components}
\title{Two-mode components}
\usage{
twomode_components(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A list including the number of components in the network,
the fragmentation of the network (number of components/number of nodes in the column nodeset),
and the component membership of each node in that nodeset.
}
\description{
This function identifies components in a two-mode network.
}
\details{
Note that this function applies only to one dimension/mode (the columns).
Use a transposed matrix to return values for the other dimension/mode.
}
\examples{
twomode_components(mat)
}
\seealso{
Other two-mode functions: 
\code{\link{twomode_2x2}()},
\code{\link{twomode_clustering}()},
\code{\link{twomode_coherence}()},
\code{\link{twomode_constraint}()},
\code{\link{twomode_dominance}()},
\code{\link{twomode_modularity}()},
\code{\link{twomode_netlm}()},
\code{\link{twomode_smallworld}()}
}
\concept{two-mode functions}
