% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_constraint}
\alias{twomode_constraint}
\title{Two-mode constraint}
\usage{
twomode_constraint(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
Constraint scores for each second-mode node
}
\description{
This function extends Ronald Burt's constraint measure to two-mode networks.
}
\details{
Note that this function returns constraint scores
for each second-mode node by default. To return constraint scores
for each first-mode node, please pass the function the transpose of the matrix. 
See Ron Burt's work on structural holes for more details.
}
\examples{
twomode_constraint(mat)
}
\seealso{
Other two-mode functions: 
\code{\link{centrality_betweenness}()},
\code{\link{centrality_closeness}()},
\code{\link{centrality_degree}()},
\code{\link{netlm2}()},
\code{\link{twomode_2x2}()},
\code{\link{twomode_clustering}()},
\code{\link{twomode_coherence}()},
\code{\link{twomode_components}()},
\code{\link{twomode_dominance}()},
\code{\link{twomode_modularity}()},
\code{\link{twomode_smallworld}()}
}
\concept{two-mode functions}
