% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_dominance}
\alias{twomode_dominance}
\title{Two-mode dominance}
\usage{
twomode_dominance(mat, attr = NULL)
}
\arguments{
\item{mat}{An affiliation or incidence matrix. For centralization around rows, simply transpose the matrix first (\code{t()})}

\item{attr}{Optionally, an attribute vector.}
}
\description{
This function allows you to calculate how (degree) centralized a two-mode graph is.
}
\examples{
twomode_dominance(mat)
twomode_dominance(mat, attr = gdp2010)
}
\seealso{
Other two-mode functions: \code{\link{twomode_clustering}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_fragmentation}},
  \code{\link{twomode_lattice}},
  \code{\link{twomode_smallworld}}
}
