% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_fragmentation}
\alias{twomode_fragmentation}
\title{Two-mode fragmentation}
\usage{
twomode_fragmentation(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A list including the number of components in the network,
the fragmentation of the network (number of components/number of nodes in the column nodeset),
and the component membership of each node in that nodeset.
}
\description{
This function identifies components in a two-mode network.
}
\details{
Note that this function applies only to one dimension/mode (the columns).
Use a transposed matrix to return values for the other dimension/mode.
}
\examples{
twomode_fragmentation(mat)
}
\seealso{
Other two-mode: \code{\link{twomode_centralization_between}},
  \code{\link{twomode_clustering}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_dominance}},
  \code{\link{twomode_lattice}}
}
