% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_modularity}
\alias{twomode_modularity}
\title{Two-mode modularity}
\usage{
twomode_modularity(mat, attr = NULL)
}
\arguments{
\item{mat}{A matrix}

\item{attr}{A vector of group assignment for the second mode node set (by default)}
}
\value{
A modularity score
}
\description{
This function calculates modularity of two-mode networks.
A vector of group assignment can be given for one of the two
node sets which will make assignment more rapid.
}
\details{
Note that this function has defaults that make sense for use with 
the gnevar datasets. Please contact me if you would like to make this function
more general.
}
\examples{
\dontrun{
twomode_modularity(mat, attr)
}
}
\seealso{
twomode_coherence

twomode_dominance

Other two-mode functions: \code{\link{twomode_2x2}},
  \code{\link{twomode_clustering}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_components}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_dominance}},
  \code{\link{twomode_netlm}},
  \code{\link{twomode_smallworld}}
}
\concept{two-mode functions}
