% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_qap.R
\name{twomode_netlm}
\alias{twomode_netlm}
\title{Two-mode quadratic assignment procedure}
\usage{
twomode_netlm(DV, IV, names, rep = 1000)
}
\arguments{
\item{DV}{A two-mode matrix object}

\item{IV}{A list of two-mode matrix objects}

\item{names}{IV names}

\item{rep}{Number of permutations. Default: 1000.}
}
\description{
This function allows you to calculate how much two-mode clustering there is.
}
\examples{
\dontrun{
twomode_netlm(matrix, list(matrix2, matrix3),
              c("Intercept", "Explainer"))
}
}
\seealso{
Other two-mode functions: 
\code{\link{twomode_2x2}()},
\code{\link{twomode_clustering}()},
\code{\link{twomode_coherence}()},
\code{\link{twomode_components}()},
\code{\link{twomode_constraint}()},
\code{\link{twomode_dominance}()},
\code{\link{twomode_modularity}()},
\code{\link{twomode_smallworld}()}
}
\concept{two-mode functions}
