% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_add.R
\name{add}
\alias{add}
\alias{join_nodes}
\alias{join_ties}
\alias{add_nodes}
\alias{add_ties}
\alias{add_node_attribute}
\alias{add_tie_attribute}
\alias{mutate_ties}
\alias{select_ties}
\alias{filter_ties}
\alias{rename_ties}
\alias{summarise_ties}
\title{Adding and copying attributes from one network to another}
\usage{
join_nodes(
  .data,
  source,
  by = NULL,
  join_type = c("full", "left", "right", "inner")
)

join_ties(object, object2, attr_name)

add_nodes(.data, nodes, attribute = NULL)

add_ties(.data, ties, attribute = NULL)

add_node_attribute(.data, attr_name, vector)

add_tie_attribute(.data, attr_name, vector)

mutate_ties(.data, ...)

select_ties(.data, ...)

filter_ties(.data, ...)

rename_ties(.data, ...)

summarise_ties(.data, ...)
}
\arguments{
\item{.data}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}

\item{object2}{A second object to copy nodes or edges from.}

\item{attr_name}{Name of the new attribute in the resulting object.}

\item{vector}{A vector of values for the new attribute.}
}
\description{
These functions allow users to add nodes, ties, or attributes to the nodes or ties
of a network.
The \verb{join_*()}, \verb{mutate_*()}, \verb{select_*()}, \verb{filter_*()}, \verb{rename_*()}, and \verb{summarise_*()}
functions adapt the \code{{dplyr}}-type syntax to work with networks of any type.
The \verb{add_*()} functions operate similarly to in \code{{igraph}}.
}
\section{Functions}{
\itemize{
\item \code{join_nodes()}: Copies node attributes from a given graph into specified graph

\item \code{join_ties()}: Copies ties from another graph to specified graph and
adds a tie attribute identifying the ties that were newly added

\item \code{add_nodes()}: Add additional ties to a network

\item \code{add_ties()}: Add additional ties to a network

\item \code{add_node_attribute()}: Insert specified values from a vector into the graph
as node attributes

\item \code{add_tie_attribute()}: Insert specified values from a vector into the network.
as tie attributes

\item \code{mutate_ties()}: Tidy way to add vector as tie attributes.

\item \code{select_ties()}: Tidy way to select tie attributes.

\item \code{filter_ties()}: Tidy way to filter ties based on a logical statement with
relation to some tie attribute.

\item \code{rename_ties()}: Tidy way to rename tie attributes.

\item \code{summarise_ties()}: Tidy way to summarise tie attributes.

}}
\examples{
  other <- create_empty(4) \%>\% mutate(name = c("Sue", "Tim", "Pam", "Mark"))
  join_nodes(ison_adolescents, other)
  autographr(mpn_elite_mex)
  both <- join_ties(mpn_elite_mex, generate_random(mpn_elite_mex), "random")
  autographr(both)
  random <- to_uniplex(both, "random")
  autographr(random)
  autographr(to_uniplex(both, "orig"))
  add_nodes(ison_adolescents, 4, list(name = c("Matthew", "Mark", "Luke", "Tim")))
  add_ties(ison_adolescents, c(1,2), list(time = 2, increment = -1))
add_node_attribute(mpn_elite_mex, "wealth", 1:35)
add_node_attribute(mpn_elite_usa_advice, "wealth", 1:14)
add_tie_attribute(ison_adolescents, "weight", c(1,2,1,1,1,3,2,2,3,1))
}
\seealso{
Other manipulations: 
\code{\link{as}()},
\code{\link{grab}},
\code{\link{na}},
\code{\link{reformat}},
\code{\link{split}()},
\code{\link{transform}()}
}
\concept{manipulations}
