% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_block.R
\name{blockmodel}
\alias{blockmodel}
\alias{print.block_model}
\alias{reduce_graph}
\title{Blockmodelling
Passing the function a list apply the algorithm to the joined matrices.}
\usage{
blockmodel(object, clusters)

\method{print}{block_model}(x, ...)

reduce_graph(blockmodel, block_labels = NULL)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix, from base R
\item edgelist, a data frame from base R or tibble from tibble
\item igraph, from the igraph package
\item network, from the network package
\item tbl_graph, from the tidygraph package
}}

\item{clusters}{the vector of cluster membership for the blockmodel}

\item{x}{An object of class "block_model"}

\item{...}{Additional arguments passed to generic print method}

\item{blockmodel}{a blockmodel object}

\item{block_labels}{A character vector manually providing labels
for the blocks in the blockmodel}

\item{block.content}{A string indicating which method to use for
calculating block content.
Options are: "density", "sum", "meanrowsum", "meancolsum",
"median", "min", "max".}
}
\description{
Blockmodelling
Passing the function a list apply the algorithm to the joined matrices.
}
\section{Functions}{
\itemize{
\item \code{reduce_graph}: creates an igraph object from the blockmodel output
}}

\examples{
blockmodel(ison_southern_women, membership)
}
