% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmodel_vis.R
\name{blockmodel_vis}
\alias{blockmodel_vis}
\alias{plot.block_model}
\alias{ggtree}
\alias{ggidentify_clusters}
\title{ggplot2-based plotting of blockmodel results}
\usage{
\method{plot}{block_model}(x, ...)

ggtree(hc, k = NULL)

ggidentify_clusters(hc, census, method = c("elbow", "strict"))
}
\arguments{
\item{x}{A blockmodel-class object.}

\item{...}{Additional arguments passed on to ggplot2.}

\item{hc}{a hierarchical cluster object}

\item{k}{number of clusters. By default NULL,
but, if specified, \code{ggtree} will color branches and
add a line to indicate where the corresponding cluster
cut would be.}

\item{census}{output from some node_*_census function}

\item{method}{only "elbow" is currently implemented.}
}
\description{
ggplot2-based plotting of blockmodel results

Plots for deciding on the number of network clusters
}
\examples{
usa_concor <- blockmodel_concor(mpn_elite_usa_advice)
plot(usa_concor)
res <- cluster_regular_equivalence(mpn_elite_mex)
ggtree(res, 3)
ggidentify_clusters(res, node_triad_census(mpn_elite_mex))
}
