% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brandes.R
\docType{data}
\name{brandes}
\alias{brandes}
\title{One-mode centrality demonstration structure}
\format{
A tidygraph \code{tbl_graph} with 11 nodes and 24 edges.
}
\usage{
data(brandes)
}
\description{
One-mode centrality demonstration structure
}
\keyword{datasets}
