% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\alias{cluster_structural_equivalence}
\alias{cluster_regular_equivalence}
\title{Clustering algorithms}
\usage{
cluster_structural_equivalence(object)

cluster_regular_equivalence(object)
}
\arguments{
\item{object}{a migraph-consistent object}
}
\description{
These functions combine an appropriate \verb{_census()} function
together with methods for calculating the hierarchical clusters
provided by a certain distance calculation.
}
\examples{
ggtree(cluster_structural_equivalence(mpn_elite_mex))
ggtree(cluster_regular_equivalence(mpn_elite_mex))
}
