% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_cohesion.R
\name{cohesion}
\alias{cohesion}
\alias{network_density}
\alias{network_components}
\alias{network_cohesion}
\alias{network_adhesion}
\alias{network_diameter}
\alias{network_length}
\title{Measures of network cohesion or connectedness}
\usage{
network_density(object)

network_components(object)

network_cohesion(object)

network_adhesion(object)

network_diameter(object)

network_length(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}
}
\description{
These functions return values or vectors relating to how connected a network is
and the number of nodes or edges to remove that would increase fragmentation.
}
\section{Functions}{
\itemize{
\item \code{network_density()}: summarises the ratio of ties
to the number of possible ties.

\item \code{network_components()}: Returns number of (strong) components in the network.
To get the 'weak' components of a directed graph,
please use \code{to_undirected()} first.

\item \code{network_cohesion()}: Returns the minimum number of nodes to remove
from the network needed to increase the number of components.

\item \code{network_adhesion()}: Returns the minimum number of edges needed
to remove from the network to increase the number of components.

\item \code{network_diameter()}: Returns the maximum path length in the network.

\item \code{network_length()}: Returns the average path length in the network.

}}
\examples{
network_density(mpn_elite_mex)
network_density(mpn_elite_usa_advice)
network_cohesion(ison_marvel_relationships)
network_cohesion(to_giant(ison_marvel_relationships))
network_adhesion(ison_marvel_relationships)
network_adhesion(to_giant(ison_marvel_relationships))
network_diameter(ison_marvel_relationships)
network_diameter(to_giant(ison_marvel_relationships))
network_length(ison_marvel_relationships)
network_length(to_giant(ison_marvel_relationships))
}
\references{
White, Douglas R and Frank Harary. 2001.
"The Cohesiveness of Blocks In Social Networks: Node Connectivity and Conditional Density."
\emph{Sociological Methodology} 31(1): 305-59.
}
\seealso{
Other measures: 
\code{\link{centralisation}},
\code{\link{centrality}},
\code{\link{closure}},
\code{\link{diversity}},
\code{\link{features}},
\code{\link{holes}},
\code{\link{tie_centrality}}
}
\concept{measures}
