% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_community.R
\name{community}
\alias{community}
\alias{node_kernighanlin}
\alias{node_walktrap}
\alias{node_edge_betweenness}
\alias{node_fast_greedy}
\title{Community graph partitioning algorithms}
\usage{
node_kernighanlin(object)

node_walktrap(object, times = 50)

node_edge_betweenness(object)

node_fast_greedy(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}
}
\description{
Community graph partitioning algorithms
}
\section{Functions}{
\itemize{
\item \code{node_kernighanlin()}: A greedy, iterative, deterministic graph
partitioning algorithm that results in a graph with two
equally-sized communities

\item \code{node_walktrap()}: The walktrap algorithm

\item \code{node_edge_betweenness()}: The edge-betweenness algorithm

\item \code{node_fast_greedy()}: The fast-greedy algorithm

}}
\examples{
node_kernighanlin(ison_adolescents)
node_kernighanlin(ison_southern_women)
node_walktrap(ison_adolescents)
node_edge_betweenness(ison_adolescents)
node_fast_greedy(ison_adolescents)
}
\references{
Kernighan, Brian W., and Shen Lin. 1970.
"An efficient heuristic procedure for partitioning graphs."
\emph{The Bell System Technical Journal} 49(2): 291-307.
\doi{10.1002/j.1538-7305.1970.tb01770.x}
}
\seealso{
Other memberships: 
\code{\link{components}()},
\code{\link{core-periphery}},
\code{\link{equivalence}}
}
\concept{memberships}
