% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_community.R
\name{community}
\alias{community}
\alias{node_kernighanlin}
\title{Community graph partitioning algorithms}
\usage{
node_kernighanlin(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix, from base R
\item edgelist, a data frame from base R or tibble from tibble
\item igraph, from the igraph package
\item network, from the network package
\item tbl_graph, from the tidygraph package
}}
}
\description{
Community graph partitioning algorithms
}
\section{Functions}{
\itemize{
\item \code{node_kernighanlin}: A greedy, iterative, deterministic graph
partitioning algorithm that results in a graph with two
equally-sized communities
}}

\examples{
node_kernighanlin(ison_adolescents)
node_kernighanlin(ison_southern_women)
}
