% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_components.R
\name{components}
\alias{components}
\alias{node_components}
\alias{node_weak_components}
\alias{node_strong_components}
\alias{node_coreness}
\title{Component partitioning algorithms}
\usage{
node_components(object)

node_weak_components(object)

node_strong_components(object)

node_coreness(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix, from base R
\item edgelist, a data frame from base R or tibble from tibble
\item igraph, from the igraph package
\item network, from the network package
\item tbl_graph, from the tidygraph package
}}
}
\description{
Component partitioning algorithms
}
\section{Functions}{
\itemize{
\item \code{node_components}: Returns nodes' component membership
using edge direction where available.

\item \code{node_weak_components}: Returns nodes' component membership
ignoring edge direction.

\item \code{node_strong_components}: Returns nodes' component membership
based on edge direction.

\item \code{node_coreness}: Returns k-cores
}}

\examples{
node_components(mpn_bristol)
node_weak_components(mpn_bristol)
node_strong_components(mpn_bristol)
node_coreness(ison_adolescents)
}
\seealso{
Other membership: 
\code{\link{community}},
\code{\link{equivalence}}
}
\concept{membership}
