% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_create.R
\name{create}
\alias{create}
\alias{create_empty}
\alias{create_filled}
\alias{create_ring}
\alias{create_star}
\alias{create_tree}
\alias{create_lattice}
\alias{create_components}
\alias{create_core}
\title{Make networks with defined structures}
\usage{
create_empty(n, directed = FALSE)

create_filled(n, directed = FALSE)

create_ring(n, directed = FALSE, width = 1, ...)

create_star(n, directed = FALSE)

create_tree(n, directed = FALSE, width = 2)

create_lattice(n, directed = FALSE, width = 8)

create_components(n, directed = FALSE, membership = NULL)

create_core(n, directed = FALSE, membership = NULL)
}
\arguments{
\item{n}{Given:
\itemize{
\item A single integer, e.g. \code{n = 10},
a one-mode network will be created.
\item A vector of two integers, e.g. \code{n = c(5,10)},
a two-mode network will be created.
\item A migraph-compatible object,
a network of the same dimensions will be created.
}}

\item{directed}{Logical whether the graph should be directed.
By default \code{directed = FALSE}.
If the opposite direction is desired, use \code{to_redirected()}.}

\item{width}{Integer specifying the width of the ring,
breadth of the branches, or maximum extent of the neighbourbood.}

\item{...}{Additional arguments passed on to \code{{igraph}}.}

\item{membership}{A vector of partition membership as integers.
If left as \code{NULL} (the default), nodes in each mode will be
assigned to two, equally sized partitions.}
}
\value{
By default an \code{igraph} object is returned,
but this can be coerced into other types of objects
using \code{as_edgelist()}, \code{as_matrix()},
\code{as_tidygraph()}, or \code{as_network()}.
}
\description{
These functions create networks with particular structural properties.
They can create either one-mode or two-mode networks.
To create a one-mode network, pass the main argument \code{n} a single integer,
indicating the number of nodes in the network.
To create a two-mode network, pass \code{n} a vector of \emph{two} integers,
where the first integer indicates the number of nodes in the first mode,
and the second integer indicates the number of nodes in the second mode.
As an alternative, an existing network can be provided to \code{n}
and the number of modes, nodes, and directedness will be inferred.

By default, all networks are created as undirected.
This can be overruled with the argument \code{directed = TRUE}.
This will return a directed network in which the arcs are
out-facing or equivalent.
This direction can be swapped using \code{to_redirected()}.
In two-mode networks, this is ignored.
}
\section{Functions}{
\itemize{
\item \code{create_empty()}: Creates an empty graph of the given dimensions.

\item \code{create_filled()}: Creates a complete graph of the given dimensions,
with every possible tie realised.

\item \code{create_ring()}: Creates a ring or chord graph of the given dimensions
that loops around is of a certain width or thickness.

\item \code{create_star()}: Creates a graph of the given dimensions
that has a maximally central node

\item \code{create_tree()}: Creates a graph of the given dimensions with successive branches.

\item \code{create_lattice()}: Creates a lattice graph of the given dimensions with ties to all neighbouring nodes

\item \code{create_components()}: Creates a graph in which the nodes are clustered
into separate components.

\item \code{create_core()}: Creates a graph with a certain proportion of nodes
being core nodes, densely tied to each other and peripheral nodes,
and the rest peripheral, tied only to the core.

}}
\section{Lattice graphs}{

\code{create_lattice()} creates both two-dimensional grid and triangular lattices with as even dimensions
as possible.
When the \code{width} parameter is set to 4, nodes cannot have (in or out) degrees larger than 4.
This creates regular square grid lattices where possible.
Such a network is bipartite, that is partitionable into two types that are not adjacent to any
of their own type.
If the number of nodes is a prime number, it will only return a chain (a single dimensional lattice).

A \code{width} parameter of 8 creates a network where the maximum degree of any nodes is 8.
This can create a triangular mesh lattice or a Queen's move lattice, depending on the dimensions.
A \code{width} parameter of 12 creates a network where the maximum degree of any nodes is 12.
Prime numbers of nodes will return a chain.
}

\examples{
autographr(create_empty(10)) + autographr(create_filled(10))
autographr(create_empty(c(8,6))) + autographr(create_filled(c(8,6)))
autographr(create_ring(8, width = 2)) +
autographr(create_ring(c(8,6), width = 2))
autographr(create_star(12)) +
autographr(create_star(12, directed = TRUE)) +
autographr(create_star(c(12,1)))
autographr(create_tree(c(7,8), directed = TRUE)) +
autographr(create_tree(15, directed = TRUE), "tree") +
autographr(create_tree(15, directed = TRUE, width = 3), "tree")
autographr(create_lattice(12, width = 4), layout = "kk") +
autographr(create_lattice(12, width = 8), layout = "kk") +
autographr(create_lattice(12, width = 12), layout = "kk")
autographr(create_components(10, membership = c(1,1,1,2,2,2,3,3,3,3)))
autographr(create_components(c(10, 12)))
autographr(create_core(6)) +
autographr(create_core(6, membership = c(1,1,1,1,2,2))) +
autographr(create_core(c(6,6)))
}
\seealso{
\link{as}

Other makes: 
\code{\link{generate}},
\code{\link{read}}
}
\concept{makes}
