% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{create_nest}
\alias{create_nest}
\title{Nested two-mode graph}
\usage{
create_nest(node1, node2)
}
\arguments{
\item{node1}{Number of nodes in the first node set}

\item{node2}{Number of nodes in the second node set}
}
\description{
Creates a nested two-mode network
}
\details{
Will construct an affiliation matrix,
with decreasing fill across node2.
TODO: Make node set matching optional.
}
\examples{
\dontrun{
create_nest(10, 12)
}
}
