% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraph-defunct.R
\name{defunct}
\alias{defunct}
\alias{cluster_triad_census}
\title{Defunct functions}
\usage{
cluster_triad_census(object, clusters)
}
\value{
\code{group_triad_census()}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Deprecated on 2021-10-26.
}
\keyword{internal}
