% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraph-defunct.R
\name{defunct}
\alias{defunct}
\alias{graph_clustering}
\alias{cluster_triad_census}
\alias{ggraphgrid}
\alias{test_cug}
\alias{graph_dimensions}
\alias{netlm}
\title{Functions that have been renamed, superseded, or are no longer working}
\usage{
graph_clustering(object)

cluster_triad_census(object, clusters)

ggraphgrid(x, algorithm = c("kk", "fr"))

test_cug(object, FUN, ..., nSim = 1000)

graph_dimensions(object)

netlm(formula, data, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\section{Functions}{
\itemize{
\item \code{graph_clustering}: Deprecated on 2021-10-18.
Returns \code{graph_transitivity()} or \code{graph_equivalency()},
depending on whether the object is a one-mode or two-mode object, respectively.

\item \code{cluster_triad_census}: Deprecated on 2021-10-26.
Returns \code{group_triad_census()}

\item \code{ggraphgrid}: Deprecated on 2021-10-26.

\item \code{test_cug}: Deprecated on 2021-11-08.
Returns \code{test_random()}

\item \code{graph_dimensions}: Deprecated on 2021-11-10.
Returns \code{graph_dims()}

\item \code{netlm}: Deprecated on 2021-11-16.
Returns \code{network_reg()}
}}

\keyword{internal}
