% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraph-defunct.R
\name{defunct}
\alias{defunct}
\alias{ggidentify}
\alias{ggdistrib}
\alias{project_rows}
\alias{project_cols}
\alias{mutate_edges}
\alias{edge_mutual}
\title{Functions that have been renamed, superseded, or are no longer working}
\usage{
ggidentify(object, node_measure, identify_function = max)

ggdistrib(object, node_measure)

project_rows(object)

project_cols(object)

mutate_edges(object, object2, attr_name)

edge_mutual(object)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\section{Functions}{
\itemize{
\item \code{ggidentify}: Deprecated on 2022-01-24.
This function is deprecated and its functionality is included in the more
general purpose \code{autographr()} function. Please refer to its documentation
for more details about the new implementation.

\item \code{ggdistrib}: Deprecated on 2022-03-23.
This function is deprecated and its functionality is included in the more
general purpose \code{autographr()} function. Please refer to its documentation
for more details about the new implementation.

\item \code{project_rows}: Deprecated on 2022-03-23.

\item \code{project_cols}: Deprecated on 2022-03-23.

\item \code{mutate_edges}: Deprecated on 2022-03-29.

\item \code{edge_mutual}: Deprecated on 2022-04-05.
}}

\keyword{internal}
