% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\alias{graph_blau_index}
\alias{graph_ei_index}
\title{Measures of network diversity}
\usage{
graph_blau_index(object, attribute, clusters = NULL)

graph_ei_index(object, attribute)
}
\arguments{
\item{object}{A data frame edgelist, matrix, igraph, tidygraph, or
network object.}

\item{attribute}{The name of a vertex attribute to measure the diversity of.}

\item{clusters}{A nodal cluster membership vector.}
}
\description{
Measures of network diversity
}
