% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{edge}
\alias{edge}
\alias{edge_mutual}
\alias{edge_multiple}
\alias{edge_loop}
\alias{edge_betweenness}
\alias{edge_closeness}
\title{Identifying edges by certain properties}
\usage{
edge_mutual(object)

edge_multiple(object)

edge_loop(object)

edge_betweenness(object)

edge_closeness(object)
}
\arguments{
\item{object}{A migraph-consistent class object
(matrix, edgelist, igraph, network, tidygraph)}
}
\description{
Identifying edges by certain properties
}
\section{Functions}{
\itemize{
\item \code{edge_mutual}: Identify edges that are mutual/reciprocated

\item \code{edge_multiple}: Identify edges that are multiples

\item \code{edge_loop}: Identify edges that are loops

\item \code{edge_betweenness}: Calculate number of shortest paths going through an edge

\item \code{edge_closeness}: Calculate the average distance of a tie to
other ties in the network. Row names indicate the node at the head of the
first tie and column names indicate the node at the tail of the last tie.
}}

\examples{
edge_mutual(ison_m182)
edge_multiple(ison_m182)
edge_loop(ison_m182)
edge_betweenness(ison_coleman)
edge_closeness(ison_coleman)
}
