% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_equivalence.R
\name{equivalence}
\alias{equivalence}
\alias{node_structural_equivalence}
\alias{node_regular_equivalence}
\alias{node_automorphic_equivalence}
\title{Equivalence clustering algorithms}
\usage{
node_structural_equivalence(
  object,
  cluster = c("hier", "concor"),
  select = c("strict", "elbow", "silhouette")
)

node_regular_equivalence(object, select = c("strict", "elbow", "silhouette"))

node_automorphic_equivalence(
  object,
  select = c("strict", "elbow", "silhouette")
)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix, from base R
\item edgelist, a data frame from base R or tibble from tibble
\item igraph, from the igraph package
\item network, from the network package
\item tbl_graph, from the tidygraph package
}}

\item{select}{Character string indicating which method
should be used to select the number of clusters to cut
the tree at.
By default "strict" to return classes with members only
when strictly equivalent.
Other options to relax this strict assumption generally
provide more useful results and include "elbow".}
}
\description{
These functions combine an appropriate \verb{_census()} function
together with methods for calculating the hierarchical clusters
provided by a certain distance calculation.

A \code{plot()} method exists for investigating the dendrogram
of the hierarchical cluster and showing the returned cluster
assignment.
}
\section{Functions}{
\itemize{
\item \code{node_structural_equivalence}: Returns nodes' membership in
structurally equivalent classes

\item \code{node_regular_equivalence}: Returns nodes' membership in
regularly equivalent classes

\item \code{node_automorphic_equivalence}: Returns nodes' membership in
automorphically equivalent classes
}}

\examples{
nse_hier <- node_structural_equivalence(ison_adolescents, 
                  cluster = "hier", select = "strict")
plot(nse_hier)
nse_conc <- node_structural_equivalence(ison_adolescents, 
                  cluster = "concor", select = "strict")
plot(nse_conc)
(nre <- node_regular_equivalence(ison_algebra, "elbow"))
plot(nre)
(nre2 <- node_regular_equivalence(ison_algebra, "silhouette"))
plot(nre2)
(nae <- node_automorphic_equivalence(mpn_elite_mex, select = "elbow"))
plot(nae)
(nae2 <- node_automorphic_equivalence(mpn_elite_usa_advice, select = "elbow"))
plot(nae2)
}
\references{
Thorndike, Robert L. 1953.
"Who Belongs in the Family?".
\emph{Psychometrika}.
18 (4): 267–276.
\doi{10.1007/BF02289263}
}
