% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_equivalence.R
\name{equivalence}
\alias{equivalence}
\alias{node_equivalence}
\alias{node_structural_equivalence}
\alias{node_regular_equivalence}
\alias{node_automorphic_equivalence}
\title{Equivalence clustering algorithms}
\source{
\url{https://github.com/aslez/concoR}
}
\usage{
node_equivalence(
  object,
  census,
  k = c("silhouette", "elbow", "strict"),
  cluster = c("hierarchical", "concor"),
  distance = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  range = 8L
)

node_structural_equivalence(
  object,
  k = c("silhouette", "elbow", "strict"),
  cluster = c("hierarchical", "concor"),
  distance = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  range = 8L
)

node_regular_equivalence(
  object,
  k = c("silhouette", "elbow", "strict"),
  cluster = c("hierarchical", "concor"),
  distance = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  range = 8L
)

node_automorphic_equivalence(
  object,
  k = c("silhouette", "elbow", "strict"),
  cluster = c("hierarchical", "concor"),
  distance = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  range = 8L
)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}

\item{census}{A matrix returned by a \verb{node_*_census()} function.}

\item{k}{Typically a character string indicating which method
should be used to select the number of clusters to return.
By default \code{"silhouette"}, other options include \code{"elbow"} and \code{"strict"}.
\code{"strict"} returns classes with members only when strictly equivalent.
\code{"silhouette"} and \code{"elbow"} select classes based on the distance between
clusters or between nodes within a cluster.
Fewer, identifiable letters, e.g. \code{"e"} for elbow, is sufficient.
Alternatively, if \code{k} is passed an integer, e.g. \code{k = 3},
then all selection routines are skipped in favour of this number of clusters.}

\item{cluster}{Character string indicating whether clusters should be
clustered hierarchically (\code{"hierarchical"}) or
through convergence of correlations (\code{"concor"}).
Fewer, identifiable letters, e.g. \code{"c"} for CONCOR, is sufficient.}

\item{distance}{Character string indicating which distance metric
to pass on to \code{stats::dist}.
By default \code{"euclidean"}, but other options include
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"}, and \code{"minkowski"}.
Fewer, identifiable letters, e.g. \code{"e"} for Euclidean, is sufficient.}

\item{range}{Integer indicating the maximum number of (k) clusters
to evaluate.
Ignored when \code{k = "strict"} or a discrete number is given for \code{k}.}
}
\description{
These functions combine an appropriate \verb{_census()} function
together with methods for calculating the hierarchical clusters
provided by a certain distance calculation.

A \code{plot()} method exists for investigating the dendrogram
of the hierarchical cluster and showing the returned cluster
assignment.
}
\section{Functions}{
\itemize{
\item \code{node_equivalence()}: Returns nodes' membership in
according to their equivalence with respective to some census/class

\item \code{node_structural_equivalence()}: Returns nodes' membership in
structurally equivalent classes

\item \code{node_regular_equivalence()}: Returns nodes' membership in
regularly equivalent classes

\item \code{node_automorphic_equivalence()}: Returns nodes' membership in
automorphically equivalent classes

}}
\examples{
\donttest{
(nse <- node_structural_equivalence(mpn_elite_usa_advice))
plot(nse)
}
\donttest{
(nre <- node_regular_equivalence(mpn_elite_usa_advice,
  cluster = "concor"))
plot(nre)
}
\donttest{
(nae <- node_automorphic_equivalence(mpn_elite_usa_advice,
  k = "elbow"))
plot(nae)
}
}
\references{
Thorndike, Robert L. 1953.
"Who Belongs in the Family?".
\emph{Psychometrika}, 18(4): 267–76.
\doi{10.1007/BF02289263}.

Rousseeuw, Peter J. 1987.
“Silhouettes: A Graphical Aid to the Interpretation and Validation of Cluster Analysis.”
\emph{Journal of Computational and Applied Mathematics}, 20: 53–65.
\doi{10.1016/0377-0427(87)90125-7}.
}
\seealso{
Other memberships: 
\code{\link{community}},
\code{\link{components}()},
\code{\link{core-periphery}}
}
\concept{memberships}
