% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\alias{generate_random}
\title{Create networks from particular probabilities}
\usage{
generate_random(n, p, m)
}
\arguments{
\item{n}{Integer of length 1 or 2.}

\item{p}{Number of edges in the network over the number of edges possible}

\item{m}{Number of edges in the network}
}
\description{
Create networks from particular probabilities
}
\details{
Creates a random network.
If \code{length(n)==1}, then a one-mode network will be returned,
equivalent to an Erdös-Renyi graph.
If \code{length(n)==1}, then a two-mode network will be returned.
The first number is the number of nodes in the first nodeset (rows),
and the second number becomes the number of nodes in the second nodeset (columns).
}
\examples{
autographr(generate_random(c(10, 12), 0.25))
}
\seealso{
Other creation: 
\code{\link{create}}
}
\concept{creation}
