% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{get_biconnected}
\alias{get_biconnected}
\title{Get biconnected subgraphs}
\usage{
get_biconnected(object)
}
\arguments{
\item{object}{}
}
\value{
a list of biconnected components in the original network
}
\description{
Get biconnected subgraphs
}
\examples{

}
