% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggevolution.R
\name{ggatyear}
\alias{ggatyear}
\title{Plotting a network at a particular timepoint (year)}
\usage{
ggatyear(edgelist, year, ...)
}
\arguments{
\item{edgelist}{A manyverse edgelist, expecting \code{Beg} and \code{End} variables,
among others}

\item{year}{Numeric year, gets expanded to first of January that year}

\item{...}{Additional arguments passed on to \code{autographr()}.}
}
\value{
A plot of the network of agreements signed in the specified year.
}
\description{
Plotting a network at a particular timepoint (year)
}
\examples{
\dontrun{
ggatyear(membs, 1900)
}
}
