% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdistrib.R
\name{ggdistrib}
\alias{ggdistrib}
\title{ggdistrib
\code{ggdistrib()} plots the distribution of a node level measure as a simple
histogram.}
\usage{
ggdistrib(object, node_measure)
}
\arguments{
\item{object}{a migraph-consistent object}

\item{node_measure}{some arbitrary function that runs on the object and
returns a numeric vector that can be used to scale the nodes}
}
\description{
ggdistrib
\code{ggdistrib()} plots the distribution of a node level measure as a simple
histogram.
}
\examples{
ggdistrib(brandes, node_degree)
ggdistrib(brandes, node_betweenness)
ggdistrib(brandes, node_closeness)
ggdistrib(brandes, node_eigenvector)
}
