% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggidentify.R
\name{ggidentify}
\alias{ggidentify}
\title{ggidentify}
\usage{
ggidentify(object, node_measure, identify_function = max)
}
\arguments{
\item{object}{a migraph-consistent object}

\item{node_measure}{some arbitrary function that runs on the object and
returns a numeric vector that can be used to scale the nodes}

\item{identify_function}{a function for the identification of a single node,
e.g. max, min, mean, etc.}
}
\description{
This function is deprecated and its functionality is included in the more
general purpose \code{autographr()} function. Please refer to its documentation
for more details about the new implementation.

Visualising graphs and identifying nodes with maximum values of the specified
measure.
}
\examples{
ggidentify(brandes, node_degree)
# ->
autographr(brandes, node_measure = node_degree, identify_function = max)
ggidentify(brandes, node_betweenness)
# ->
autographr(brandes, node_measure = node_betweenness, identify_function = max)
ggidentify(brandes, node_closeness)
# ->
autographr(brandes, node_measure = node_closeness, identify_function = max)
ggidentify(brandes, node_eigenvector)
# ->
autographr(brandes, node_measure = node_eigenvector, identify_function = max)
}
\keyword{internal}
