% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\alias{node_names}
\alias{node_attribute}
\alias{edge_attribute}
\alias{edge_weights}
\alias{graph_nodes}
\alias{graph_edges}
\alias{graph_dims}
\alias{graph_node_attributes}
\alias{graph_edge_attributes}
\title{Helpers to grab various attributes from nodes or edges in a graph}
\usage{
node_names(object)

node_attribute(object, attribute)

edge_attribute(object, attribute)

edge_weights(object)

graph_nodes(object)

graph_edges(object)

graph_dims(object)

graph_node_attributes(object)

graph_edge_attributes(object)
}
\arguments{
\item{object}{A data frame edgelist, matrix, igraph, tidygraph, or
network object.}

\item{attribute}{An attribute name.}
}
\description{
Helpers to grab various attributes from nodes or edges in a graph
}
