% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{graph_balance}
\alias{graph_balance}
\title{Structural balance}
\source{
Gábor Csárdi: http://r.789695.n4.nabble.com/Social-Network-Analysis-td825041.html
}
\usage{
graph_balance(object)
}
\arguments{
\item{object}{a migraph-consistent object}
}
\description{
Structural balance
}
