% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{graph_balance}
\alias{graph_balance}
\title{Structural balance}
\source{
\code{{signnet}} by David Schoch
}
\usage{
graph_balance(object, method = "triangles")
}
\arguments{
\item{object}{a migraph-consistent object}

\item{method}{one of "triangles" (the default), "walk", or "frustration".}
}
\value{
"triangles" returns the proportion of balanced triangles,
ranging between \code{0} if all triangles are imbalanced and \code{1} if all triangles are balanced.
}
\description{
Structural balance
}
