% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{graph_census}
\alias{graph_census}
\alias{graph_dyad_census}
\alias{graph_triad_census}
\title{Censuses for the whole graph}
\usage{
graph_dyad_census(object)

graph_triad_census(object)
}
\arguments{
\item{object}{a migraph-consistent object}
}
\description{
Censuses for the whole graph
}
\examples{
graph_dyad_census(ison_coleman)
graph_triad_census(ison_coleman)
}
