% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{graph_census}
\alias{graph_census}
\alias{graph_mixed_census}
\alias{graph_dyad_census}
\alias{graph_triad_census}
\title{Censuses for the whole graph}
\source{
Alejandro Espinosa 'netmem'
}
\usage{
graph_mixed_census(object, object2)

graph_dyad_census(object)

graph_triad_census(object)
}
\arguments{
\item{object}{A migraph-consistent object.}

\item{object2}{A second, two-mode migraph-consistent object.}
}
\description{
Censuses for the whole graph
}
\examples{
marvel_friends <- to_unsigned(ison_marvel_relationships, "positive")
(mixed_cen <- graph_mixed_census(marvel_friends, ison_marvel_teams))
graph_dyad_census(ison_adolescents)
graph_triad_census(ison_adolescents)
}
\references{
Hollway, James, Alessandro Lomi, Francesca Pallotti, and Christoph Stadtfeld.
“\doi{10.1017/nws.2017.8}{Multilevel Social Spaces: The Network Dynamics of Organizational Fields}.”
\emph{Network Science} 5, no. 2 (June 2017): 187–212.
}
