% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is}
\alias{is}
\alias{is_twomode}
\alias{is_weighted}
\alias{is_directed}
\alias{is_labelled}
\title{Tests of network properties}
\usage{
is_twomode(object)

is_weighted(object)

is_directed(object)

is_labelled(object)
}
\arguments{
\item{object}{A migraph-consistent class object (matrix, edgelist, igraph, network, tidygraph)}
}
\value{
TRUE if object is a two-mode network, otherwise FALSE

TRUE if object is a weighted network, otherwise FALSE

TRUE if object is a directed network, otherwise FALSE

TRUE if object is a labelled network, otherwise FALSE
}
\description{
Tests of network properties
}
\examples{
is_twomode(southern_women)
is_weighted(southern_women)
is_directed(southern_women)
is_labelled(southern_women)
}
