% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is}
\alias{is}
\alias{is_twomode}
\alias{is_weighted}
\alias{is_directed}
\alias{is_labelled}
\alias{is_signed}
\alias{is_connected}
\title{Tests of network properties}
\usage{
is_twomode(object)

is_weighted(object)

is_directed(object)

is_labelled(object)

is_signed(object)

is_connected(object, method = c("weak", "strong"))
}
\arguments{
\item{object}{A migraph-consistent class object
(matrix, edgelist, igraph, network, tidygraph)}

\item{method}{Whether to identify components if only "weak"ly connected
or also "strong"ly connected.}
}
\value{
TRUE if object is a two-mode network, otherwise FALSE

TRUE if object is a weighted network, otherwise FALSE

TRUE if object is a directed network, otherwise FALSE

TRUE if object is a labelled network, otherwise FALSE

TRUE if object is a signed network, otherwise FALSE

TRUE if object is a connected network, otherwise FALSE
}
\description{
Tests of network properties
}
\examples{
is_twomode(southern_women)
is_weighted(southern_women)
is_directed(southern_women)
is_labelled(southern_women)
is_signed(southern_women)
is_connected(southern_women)
}
\seealso{
Other manipulation: 
\code{\link{coercion}},
\code{\link{project}},
\code{\link{to}}
}
\concept{manipulation}
