% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ison.R
\docType{data}
\name{ison_southern_women}
\alias{ison_southern_women}
\title{Two-mode southern women (Davis, Gardner and Gardner 1941)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> IGRAPH f8d9f5f UN-B 32 93 -- 
#> + attr: type (v/l), name (v/c)
#> + edges from f8d9f5f (vertex names):
#>  [1] EVELYN   --E1 EVELYN   --E2 EVELYN   --E3 EVELYN   --E4 EVELYN   --E5
#>  [6] EVELYN   --E6 EVELYN   --E8 EVELYN   --E9 LAURA    --E1 LAURA    --E2
#> [11] LAURA    --E3 LAURA    --E5 LAURA    --E6 LAURA    --E7 LAURA    --E8
#> [16] THERESA  --E2 THERESA  --E3 THERESA  --E4 THERESA  --E5 THERESA  --E6
#> [21] THERESA  --E7 THERESA  --E8 THERESA  --E9 BRENDA   --E1 BRENDA   --E3
#> [26] BRENDA   --E4 BRENDA   --E5 BRENDA   --E6 BRENDA   --E7 BRENDA   --E8
#> [31] CHARLOTTE--E3 CHARLOTTE--E4 CHARLOTTE--E5 CHARLOTTE--E7 FRANCES  --E3
#> [36] FRANCES  --E5 FRANCES  --E6 FRANCES  --E8 ELEANOR  --E5 ELEANOR  --E6
#> + ... omitted several edges
}\if{html}{\out{</div>}}
}
\usage{
data(ison_southern_women)
}
\description{
Two-mode network dataset collected by Davis, Gardner and Gardner (1941)
about the attendance pattern of women at informal social events
during a 9 month period. Events and women are named.
}
\references{
Davis, Allison, Burleigh B. Gardner, and Mary R. Gardner. 1941.
\emph{Deep South}.
Chicago: University of Chicago Press.
}
\keyword{datasets}
