% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_edges.R, R/measure_connected.R
\name{mark}
\alias{mark}
\alias{edge_multiple}
\alias{edge_loop}
\alias{edge_reciprocal}
\alias{edge_bridges}
\alias{node_cuts}
\title{Identify nodes and edges based on some criteria}
\usage{
edge_multiple(object)

edge_loop(object)

edge_reciprocal(object)

edge_bridges(object)

node_cuts(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}
}
\description{
These functions return logical vectors of the length of the
nodes or edges in a network that identify nodes or edges that
meet some criteria.
Such functions are often useful for identifying particular
nodes or edges that are particularly well- or poorly-connected.
}
\section{Functions}{
\itemize{
\item \code{edge_multiple}: Returns logical of which edges are multiples

\item \code{edge_loop}: Returns logical of which edges are loops

\item \code{edge_reciprocal}: Returns logical of which edges
are mutual/reciprocated

\item \code{edge_bridges}: Returns logical of which nodes cut
or act as articulation points in a network.

\item \code{node_cuts}: Returns logical of which nodes cut
or act as articulation points in a network,
increasing the number of connected components in a graph when removed.
}}

\examples{
edge_multiple(ison_marvel_relationships)
edge_loop(ison_marvel_relationships)
edge_reciprocal(ison_algebra)
edge_bridges(ison_brandes)
node_cuts(ison_brandes)
}
