% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_edges.R, R/measure_connected.R
\name{mark}
\alias{mark}
\alias{tie_is_multiple}
\alias{tie_is_loop}
\alias{tie_is_reciprocal}
\alias{tie_is_bridge}
\alias{node_cuts}
\title{Logical whether nodes or ties meet some criteria}
\usage{
tie_is_multiple(object)

tie_is_loop(object)

tie_is_reciprocal(object)

tie_is_bridge(object)

node_cuts(object)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}
}
\description{
These functions return logical vectors of the length of the
nodes or ties in a network identifying which meet some criteria.
Such functions are often useful for identifying nodes or ties
that are particularly well- or poorly-connected.
}
\section{Functions}{
\itemize{
\item \code{tie_is_multiple}: Returns logical of which ties are multiples

\item \code{tie_is_loop}: Returns logical of which ties are loops

\item \code{tie_is_reciprocal}: Returns logical of which ties
are mutual/reciprocated

\item \code{tie_is_bridge}: Returns logical of which ties cut
or act as articulation points in a network.

\item \code{node_cuts}: Returns logical of which nodes cut
or act as articulation points in a network,
increasing the number of connected components in a graph when removed.
}}

\examples{
tie_is_multiple(ison_marvel_relationships)
tie_is_loop(ison_marvel_relationships)
tie_is_reciprocal(ison_algebra)
tie_is_bridge(ison_brandes)
node_cuts(ison_brandes)
}
\seealso{
Other marks: 
\code{\link{is}()}
}
\concept{marks}
