% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_m182.R
\docType{data}
\name{marvel}
\alias{marvel}
\alias{ison_m182}
\title{Multiplex igraph of friends, social, and task ties between 16 anonymous students}
\format{
Multiplex tidygraph of friends, social, and task ties between 16 anonymous students
}
\source{
See \code{data(studentnets.M182, package = "NetData")}
}
\usage{
data(ison_m182)
}
\description{
Multiplex igraph of friends, social, and task ties between 16 anonymous students
}
\keyword{datasets}
