% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bristol.R
\docType{data}
\name{mpn_bristol}
\alias{mpn_bristol}
\title{Bristol protest events, 1990-2002}
\format{
A matrix with 264 rows and columns. Node IDs are prefaced with a type identifier:
\describe{
  \item{1_}{150 Individuals, anonymised}
  \item{2_}{97 Bristol Civic Organizations}
  \item{3_}{17 Major Protest and Civic Events in Bristol, 1990-2002}
}
}
\source{
Knoke, Diani, Hollway, and Christopoulos. 2021. \emph{Multimodal Political Networks}. Cambridge University Press: Cambridge.
}
\usage{
mpn_bristol
}
\description{
A multimodal (3) matrix containing individuals affiliations to civic organizations
in Bristol and their participation in major protest and civic events between 1990-2002,
and the involvement of the organizations in these events.
}
\keyword{datasets}
