% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mpn.R
\docType{data}
\name{mpn_bristol}
\alias{mpn_bristol}
\title{Multimodal (3) Bristol protest events, 1990-2002 (Diani and Bison 2004)}
\format{
A \code{tbl_graph} object with 264 nodes and 1496 edges.
While represented as a two-mode network, it contains three levels:
\describe{
\item{1}{150 Individuals, anonymised}
\item{2}{97 Bristol civic organizations}
\item{3}{17 Major protest and civic events in Bristol, 1990-2002}
}
The network is named.
}
\source{
Knoke, David, Mario Diani, James Hollway, and Dimitris C Christopoulos. 2021.
\href{https://www.cambridge.org/core/books/multimodal-political-networks/43EE8C192A1B0DCD65B4D9B9A7842128}{\emph{Multimodal Political Networks}}.
Cambridge University Press. Cambridge University Press.
}
\usage{
data(mpn_bristol)
}
\description{
A multimodal (3) network containing individuals affiliations to civic organizations
in Bristol and their participation in major protest and civic events between 1990-2002,
and the involvement of the organizations in these events.
}
\references{
Diani, Mario, and Ivano Bison. 2004.
“Organizations, Coalitions, and Movements.”
\emph{Theory and Society} 33(3–4):281–309.
}
\keyword{datasets}
