% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_diffusion.R
\name{net_diffusion}
\alias{net_diffusion}
\alias{network_transmissibility}
\alias{network_infection_length}
\alias{network_reproduction}
\alias{network_immunity}
\title{Diffusion metrics for networks}
\usage{
network_transmissibility(diff_model)

network_infection_length(diff_model)

network_reproduction(diff_model)

network_immunity(diff_model)
}
\arguments{
\item{diff_model}{A valid network diffusion model,
as created by \code{as_diffusion()} or \code{play_diffusion()}.}
}
\description{
These functions allow measurement of various features of
a diffusion process:
\itemize{
\item \code{network_transmissibility()}: Measures the average transmissibility observed
in a diffusion simulation, or the number of new infections over
the number of susceptible nodes
\item \code{network_infection_length()}: Measures the average length nodes remain
infected in a compartmental model with recovery for the network as a whole
\item \code{network_reproduction()}: Measures the observed reproductive number
in a diffusion simulation as the network's transmissibility over
the network's average infection length
\item \code{network_immunity}: Measures the proportion of nodes that would need
to be protected through vaccination, isolation, or recovery for herd immunity to be reached
}
}
\section{Infection length}{

\code{network_infection_length()} measures the average length of time that
nodes in a network remain infected.
Note that in a diffusion model without recovery, average infection length
will be infinite.
This will also be the case where there is right censoring.
}

\section{Reproduction number}{

\code{network_reproduction()} measures a given diffusion's reproductive number.
Here it is calculated as:
\deqn{R = \min\left(\frac{T}{1/IL}, \bar{k}\right)}
where \eqn{T} is the observed transmissibility in a diffusion
and \eqn{IL} is the observed infection length in a diffusion.
Since \eqn{IL} can be infinite where there is no recovery
or there is right censoring,
and since network structure places an upper limit on how many
nodes each node may further infect (their degree),
this function returns the minimum of \eqn{R_0}
and the network's average degree.

Interpretation of the reproduction number is oriented around R = 1.
Where \eqn{R > 1}, the 'disease' will 'infect' more and more
nodes in the network.
Where \eqn{R < 1}, the 'disease' will not sustain itself and eventually
die out.
Where \eqn{R = 1}, the 'disease' will continue as endemic,
if conditions allow.
}

\section{Herd immunity}{

\code{network_immunity()} estimates the proportion of a network
that need to be protected from infection for herd immunity
to be achieved.
This is known as the Herd Immunity Threshold or HIT:
\deqn{1 - \frac{1}{R}}
where \eqn{R} is the reproduction number from \code{network_reproduction()}.
The HIT indicates the threshold at which
the reduction of susceptible members of the network means
that infections will no longer keep increasing.
Note that there may still be more infections after this threshold has been reached,
but there should be fewer and fewer.
These excess infections are called the \emph{overshoot}.
This function does \emph{not} take into account the structure
of the network, instead using the average degree.

Interpretation is quite straightforward.
A HIT or immunity score of 0.75 would mean that 75\% of the nodes in the network
would need to be vaccinated or otherwise protected to achieve her immunity.
To identify how many nodes this would be, multiply this proportion with the number
of nodes in the network.
}

\examples{
  smeg <- manynet::generate_smallworld(15, 0.025)
  smeg_diff <- play_diffusion(smeg, recovery = 0.2)
  plot(smeg_diff)
  autographs(smeg_diff)
  # autographd(smeg_diff)
  summary(node_adoption_time(smeg_diff), membership = adopts)
  summary(node_thresholds(smeg_diff), membership = adopts)
  summary(node_infection_length(smeg_diff))
  network_infection_length(smeg_diff)
  network_transmissibility(smeg_diff)
  # To calculate the average infection length for a given diffusion model
  network_infection_length(smeg_diff)
  # To calculate the reproduction number for a given diffusion model
  network_reproduction(smeg_diff)
  # Calculating the proportion required to achieve herd immunity
  network_immunity(smeg_diff)
  # To find the number of nodes to be vaccinated
  ceiling(network_immunity(smeg_diff) * manynet::network_nodes(smeg))
}
\references{
Kermack, W. and McKendrick, A., 1927. "A contribution to the mathematical theory of epidemics".
\emph{Proc. R. Soc. London A} 115: 700-721.
}
\seealso{
Other measures: 
\code{\link{between_centrality}},
\code{\link{close_centrality}},
\code{\link{closure}},
\code{\link{cohesion}()},
\code{\link{degree_centrality}},
\code{\link{eigenv_centrality}},
\code{\link{features}},
\code{\link{heterogeneity}},
\code{\link{hierarchy}},
\code{\link{holes}},
\code{\link{node_diffusion}}
}
\concept{measures}
