% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_play.R
\name{play_diffusion}
\alias{play_diffusion}
\title{Functions to play games on networks}
\usage{
play_diffusion(object, seeds = 1:2, thresholds = 1, steps)
}
\arguments{
\item{object}{An object of a migraph-consistent class:
\itemize{
\item matrix (adjacency or incidence) from \code{{base}} R
\item edgelist, a data frame from \code{{base}} R or tibble from \code{{tibble}}
\item igraph, from the \code{{igraph}} package
\item network, from the \code{{network}} package
\item tbl_graph, from the \code{{tidygraph}} package
}}

\item{seeds}{A valid mark vector the length of the
number of nodes in the network.}

\item{thresholds}{A numeric vector indicating the thresholds
each node has. By default 1.}

\item{steps}{The number of steps forward in the diffusion to play.
By default the number of nodes in the network.}
}
\description{
Functions to play games on networks
}
\examples{
play_diffusion(generate_smallworld(15, 0.025))
}
