% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_2x2}
\alias{plot_2x2}
\title{Plotting two-by-two plots}
\usage{
plot_2x2(dat, colour = "red")
}
\arguments{
\item{dat}{A data.frame}
}
\value{
A plot
}
\description{
Better default plotting of values traced through a two-by-two than ggplot2
}
\details{
Makes for clearer communication of structure than ggplot2 defaults
}
\examples{
\dontrun{
library(gnevar)
library(wbstats)
mil_data <- wb(country = unique(stat_actor$StatID), indicator = "MS.MIL.XPND.CN", startdate = 1960, enddate = 2018)
ally_topo <- twomode_2x2(stat_actor, ally_agree, ally_membs, mil_data, 1960, 2018)
plot_2x2(ally_topo)
}
}
