% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_conversion.R
\name{project_list}
\alias{project_list}
\title{Transform a two-mode event list into a one-mode event list}
\usage{
project_list(eventlist)
}
\arguments{
\item{eventlist}{a two-mode event list}
}
\value{
a one-mode event list
}
\description{
This function relies on `tidyverse` functions to quickly and efficiently
project a two-mode event list into a one-mode event list of nodes from 
the first mode that share connections to nodes in the second mode.
}
