% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_models.R, R/reexports_classes.R,
%   R/reexports_ggplot2.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{\%>\%}
\alias{ggplot}
\alias{ggtitle}
\alias{guides}
\alias{labs}
\alias{xlab}
\alias{ylab}
\alias{aes}
\alias{ggsave}
\alias{scale_y_discrete}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{ggsave}}, \code{\link[ggplot2:labs]{ggtitle}}, \code{\link[ggplot2]{guides}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2:scale_discrete]{scale_y_discrete}}, \code{\link[ggplot2:labs]{xlab}}, \code{\link[ggplot2:labs]{ylab}}}
}}

