% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports_ggplot2.R, R/reexports_igraph.R,
%   R/reexports_network.R, R/reexports_tidygraph.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot}
\alias{ggtitle}
\alias{labs}
\alias{xlab}
\alias{ylab}
\alias{aes}
\alias{ggsave}
\alias{grid.arrange}
\alias{is.igraph}
\alias{is_bipartite}
\alias{is.network}
\alias{as.network}
\alias{\%>\%}
\alias{with_graph}
\alias{is.tbl_graph}
\alias{.G}
\alias{.N}
\alias{.E}
\alias{mutate}
\alias{rename}
\alias{filter}
\alias{activate}
\alias{bind_edges}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{ggsave}}, \code{\link[ggplot2:labs]{ggtitle}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2:labs]{xlab}}, \code{\link[ggplot2:labs]{ylab}}}

  \item{gridExtra}{\code{\link[gridExtra:arrangeGrob]{grid.arrange}}}

  \item{igraph}{\code{\link[igraph:is_igraph]{is.igraph}}, \code{\link[igraph:make_bipartite_graph]{is_bipartite}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{network}{\code{\link[network:network]{as.network}}, \code{\link[network:network]{is.network}}}

  \item{tidygraph}{\code{\link[tidygraph:context_accessors]{.E}}, \code{\link[tidygraph:context_accessors]{.G}}, \code{\link[tidygraph:context_accessors]{.N}}, \code{\link[tidygraph]{activate}}, \code{\link[tidygraph:bind_graphs]{bind_edges}}, \code{\link[tidygraph:reexports]{filter}}, \code{\link[tidygraph:tbl_graph]{is.tbl_graph}}, \code{\link[tidygraph:reexports]{mutate}}, \code{\link[tidygraph:reexports]{rename}}, \code{\link[tidygraph]{with_graph}}}
}}

