% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uncateg.R
\docType{data}
\name{southern_women}
\alias{southern_women}
\title{Two-mode southern women dataset}
\format{
igraph graph object
}
\usage{
data(southern_women)
}
\description{
Two-mode network dataset collected by Davis, Gardner and Gardner (1941) about women and social events.
}
\references{
Davis, A., Gardner, B., and Gardner, R. 1941. \emph{Deep South}.
Chicago: University of Chicago Press.
}
\keyword{datasets}
