% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to}
\alias{to}
\alias{to_unweighted}
\alias{to_unnamed}
\alias{to_undirected}
\title{Tools for reformatting networks, graphs, and matrices}
\usage{
to_unweighted(object, threshold = 1)

to_unnamed(object)

to_undirected(object)
}
\arguments{
\item{object}{A matrix, \code{{igraph}} graph, \code{{tidygraph}} tbl_graph, or \code{{network}} object.}

\item{threshold}{For a matrix, the threshold to binarise/dichotomise at.}
}
\description{
Tools for reformatting networks, graphs, and matrices
}
\examples{
to_unweighted(project_rows(southern_women))
to_unnamed(project_rows(southern_women))
to_undirected(adolescent_society)
}
