% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_centralization_between}
\alias{twomode_centralization_between}
\title{#' Two-mode degree centralization
#'
#' This function allows you to calculate how (degree) centralized a two-mode graph is.
#' @param graph An igraph graph
#' @keywords two-mode
#' @export
#' @examples
#' twomode_centralization_degree(graph)
twomode_centralization_degree <- function(graph){
  require(igraph)
  nodeset <- names(which(igraph::degree(graph)==max(igraph::degree(graph)))) %in% 
    V(graph)$name[V(graph)$type==T]
  m <- length(which(V(graph)$type==T))
  n <- length(which(V(graph)$type!=T))
  
  sum(max(igraph::degree(graph)[which(V(graph)$type==nodeset)], na.rm=T)-
        igraph::degree(graph)[which(V(graph)$type==nodeset)])/
    ((n-1)*(m-1))
}
Two-mode betweenness centralization}
\usage{
twomode_centralization_between(graph)
}
\arguments{
\item{graph}{An igraph graph}
}
\description{
This function allows you to calculate how (betweenness) centralized a two-mode graph is.
}
\examples{
twomode_centralization_between(graph)
}
\keyword{two-mode}
