% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_centralization_degree}
\alias{twomode_centralization_degree}
\title{Two-mode degree centralization}
\usage{
twomode_centralization_degree(graph)
}
\arguments{
\item{graph}{An igraph graph}
}
\description{
This function allows you to calculate how (degree) centralized a two-mode graph is.
}
\examples{
twomode_centralization_degree(graph)
}
\references{
Borgatti, Stephen P, and Daniel S Halgin. 2011. ``Analyzing Affiliation Networks." In The SAGE Handbook of Social Network Analysis, edited by John Scott and Peter J Carrington, 417–33. London, UK: Sage.
}
