% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_centralization_degree}
\alias{twomode_centralization_degree}
\title{Two-mode degree centralization}
\usage{
twomode_centralization_degree(mat, attr = NULL)
}
\arguments{
\item{mat}{An affiliation or incidence matrix. For centralization around rows, simply transpose the matrix first (\code{t()})}

\item{attr}{Optionally, an attribute vector.}
}
\description{
This function allows you to calculate how (degree) centralized a two-mode graph is.
}
\examples{
twomode_centralization_degree(mat)
twomode_centralization_degree(mat, attr = gdp2010)
}
\seealso{
Other two-mode: \code{\link{twomode_centralization_between}},
  \code{\link{twomode_clustering}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_lattice}}
}
\keyword{two-mode}
