% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_clustering}
\alias{twomode_clustering}
\title{Two-mode clustering}
\usage{
twomode_clustering(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
This function allows you to calculate how much two-mode clustering there is.
}
\examples{
twomode.clustering(matrix)
}
\keyword{two-mode}
