% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_clustering}
\alias{twomode_clustering}
\title{Two-mode clustering}
\usage{
twomode_clustering(mat)
}
\arguments{
\item{mat}{A matrix}
}
\description{
This function allows you to calculate how much two-mode clustering there is.
}
\examples{
twomode_clustering(matrix)
}
\seealso{
Other two-mode functions: \code{\link{twomode_2x2}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_components}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_dominance}},
  \code{\link{twomode_lattice}},
  \code{\link{twomode_modularity}},
  \code{\link{twomode_smallworld}}
}
\concept{two-mode functions}
