% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_coherence}
\alias{twomode_coherence}
\title{Two-mode coherence}
\usage{
twomode_coherence(mat, attr = NULL)
}
\arguments{
\item{mat}{A matrix}
}
\value{
Average coherence across the components of the network
}
\description{
This function calculates coherence for a two-mode network
}
\details{
Note that this function applies only to one dimension/mode (the columns).
Use a transposed matrix to return values for the other dimension/mode.
}
\examples{
twomode_coherence(mat, attr)
}
\seealso{
twomode_fragmentation

Other two-mode functions: 
\code{\link{centrality_betweenness}()},
\code{\link{centrality_closeness}()},
\code{\link{centrality_degree}()},
\code{\link{netlm2}()},
\code{\link{twomode_2x2}()},
\code{\link{twomode_clustering}()},
\code{\link{twomode_components}()},
\code{\link{twomode_constraint}()},
\code{\link{twomode_dominance}()},
\code{\link{twomode_modularity}()},
\code{\link{twomode_smallworld}()}
}
\concept{two-mode functions}
