% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_constraint}
\alias{twomode_constraint}
\title{Two-mode constraint}
\usage{
twomode_constraint(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
Constraint scores for each second-mode node
}
\description{
This function extends Ronald Burt's constraint measure
to two-mode networks. Note that this function returns constraint scores
for each second-mode node by default. To return constraint scores
for each first-mode node, please pass the function the transpose of the matrix.
of
}
\details{
See Ron Burt's work on structural holes for more details
}
\examples{
twomode_constraint(mat)
}
