% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomode.R
\name{twomode_lattice}
\alias{twomode_lattice}
\title{Two-mode lattice}
\usage{
twomode_lattice(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
This function allows you to express your love of lattices.
}
\examples{
twomode_lattice(matrix)
}
\seealso{
Other two-mode functions: \code{\link{twomode_2x2}},
  \code{\link{twomode_clustering}},
  \code{\link{twomode_coherence}},
  \code{\link{twomode_constraint}},
  \code{\link{twomode_dominance}},
  \code{\link{twomode_fragmentation}},
  \code{\link{twomode_smallworld}}
}
